% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ab_info_to_fcs.R
\name{ab_info_to_fcs}
\alias{ab_info_to_fcs}
\title{Title}
\usage{
ab_info_to_fcs(
  sampledescription,
  FileNames,
  FCS.file.folder,
  channel_conjugate_match_file = system.file("extdata",
    "channel_conjugate_matches.xlsx", package = "fcexpr"),
  AbCalcFile_col = "AbCalcFile",
  AbCalcSheet_col = "AbCalcSheet",
  conjugate_to_desc = T,
  other_keywords = c("Isotype", "Clone", "totalDF", "Vendor", "Cat", "Lot"),
  clear_previous = T
)
}
\arguments{
\item{clear_previous}{}
}
\value{

}
\description{
Title
}
