% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ab_info_to_panel.R
\name{ab_info_to_panel}
\alias{ab_info_to_panel}
\title{Complement the antibody panel table with additional information from your antibody list}
\usage{
ab_info_to_panel(
  panel_file,
  panel_sheet = "Panel",
 
    antibody_list = "/Volumes/AG_Hiepe/_AG-HIEPE_Common/Antibody_List/20200705_antibody_list.xlsx",
  antibody_list_sheet = 1,
  antibody_list_cols = c("Reactivity", "Isotype", "Clone", "Vendor", "Cat",
    "Expiry.date", "Concentration.ug.ml", "Recomm.dilution")
)
}
\arguments{
\item{panel_file}{character, path to the xlsx file in which the antibody panel was calculated}

\item{panel_sheet}{character, the name of the sheet in the panel_file}

\item{antibody_list}{character, path to the xlsx file representing the antibody list with additional information for the used antibodies}

\item{antibody_list_sheet}{character, the name of the sheet in the antibody_list}

\item{antibody_list_cols}{character vector, which columns from antibody_list to append to the panel_sheet in panel_file}
}
\value{
No return value, but an appended panel_file written to disk.
}
\description{
This function matches the antibodies in a panel to those listed in an antibody list.
The antibody list may contain additional information like clone or catalog number which are to
be added to the panel table for proper annotation of the experiment.
The panel_file should be a xlsx-file as found in the Protocols-folder when running fcexpr::new_exp().
Necessary columns in panel_file and antibody_list used for matching are Antigen, Conjugate, Box, Lot.
If they do not exist this function cannot be used, yet. Columns added to the panel_file will appear just right to
the LiveDeadMarker column. Anything written there will be overwritten.
}
