% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{check_ws}
\alias{check_ws}
\title{Validate and Read a Workspace File}
\usage{
check_ws(ws)
}
\arguments{
\item{ws}{A single character string containing the path to a \code{.wsp} or
\code{.wsp.gz} file, or an object of class \link[xml2:xml_document]{xml2::xml_document}.}
}
\value{
An object of class \link[xml2:xml_document]{xml2::xml_document}.
}
\description{
Validates a workspace supplied either as an XML document or as the path to
a workspace file. File paths are read and returned as XML documents.
}
\examples{
workspace <- xml2::read_xml("<workspace/>")
check_ws(workspace)

\dontrun{
check_ws("path/to/workspace.wsp")
}
}
