% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coalesce_join.R
\name{coalesce_join}
\alias{coalesce_join}
\title{Coalescing join function which will update NA values in the left-hand data frame}
\usage{
coalesce_join(
  x,
  y,
  by = NULL,
  suffix = c(".x", ".y"),
  join = dplyr::left_join,
  ...
)
}
\arguments{
\item{x}{left-hand data frame}

\item{y}{right-hand data frame}

\item{by}{which column(s) to join by}

\item{...}{}
}
\description{
This is a combination of join functions and the coalesce function from dplyr.
It is a convenient way to solve the generic task of updating a data frame (replacing NAs)
with another one that holds additional information. In its current form the information (non-NA)
in the left-hand data frame (x) will be prioritized over that information in the right-hand
data frame (y).
}
\details{
originally from: https://alistaire.rbind.io/blog/coalescing-joins/
}
