% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/col_pal.R
\name{col_pal}
\alias{col_pal}
\title{Color palettes}
\usage{
col_pal(name = "custom", n = 100, nbrew = NULL, reverse = F)
}
\arguments{
\item{name}{name of the palette}

\item{n}{number of colors to return; may not work for every palette}

\item{nbrew}{number of color from brewer palettes}

\item{reverse}{reverse the palette}
}
\value{
a color palette as character vector
}
\description{
Common interface to get color palettes from different packages:
viridisLite, colorRamps,
\href{"https://github.com/BlakeRMills/MetBrewer"}{MetBrewer},
href{"https://github.com/karthik/wesanderson"}{wesanderson}
and RColorBrewer.
Moreover three custom palettes: "custom", "dutch" and "spanish".
}
\details{
See available palettes:
c(rev(ls("package:viridisLite"))\link{-c(1,2)},
rev(ls("package:colorRamps"))\link{-c(2,3)},
rownames(RColorBrewer::brewer.pal.info),
names(wesanderson::wes_palettes), "custom", "dutch", "spanish",
MetBrewer::MetPalettes)
Use scales::show_col() to plot color grid
}
\examples{
\dontrun{
}
}
