% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dr_to_fcs.R
\name{dr_to_fcs}
\alias{dr_to_fcs}
\title{Title}
\usage{
dr_to_fcs(
  ff.list,
  channels = NULL,
  timeChannel = "Time",
  add.sample.info,
  scale.whole = "z.score",
  scale.samples = "none",
  run.harmony = F,
  harmony.subsets,
  run.pca = F,
  run.umap = T,
  run.tsne = F,
  run.louvain = T,
  run.kmeans = F,
  run.leiden = F,
  run.hclust = F,
  run.flowClust = F,
  run.MUDAN = F,
  MUDAN.k = 30,
  n.pca.dims = 0,
  umap.n_neighbors = 15,
  umap.n_epochs = 300,
  umap.metric = "cosine",
  tsne.theta = 0.1,
  louvain_cluster_resolutions = c(seq(0.3, 0.5, 0.1)),
  leiden_cluster_resolutions = c(seq(0.3, 0.5, 0.1)),
  n_cluster = c(5:10),
  extra.cols,
  mc.cores = 1,
  save.to.disk = c("fcs", "rds"),
  save.path = file.path(getwd(), paste0(substr(gsub("\\\\.", "",
    make.names(as.character(Sys.time()))), 2, 15), "_FCS_dr")),
  exclude.extra.channels = "FSC|SSC|Time|cluster.id",
  write.scaled.channels.to.FCS = T,
  check.channels = T,
  umap.ret_model = F,
  ...
)
}
\arguments{
\item{add.sample.info}{named list of additional channels to identify samples or group them in flowjo;
e.g.: add.sample.info = list(condition = c(1,2,3,1,2,3,1,2,3), donor = c(1,1,1,2,2,2,3,3,3))}

\item{...}{}
}
\value{

}
\description{
logicle trans: 2006 - Parks_A New Logicle Display Method Avoids Deceptive Effects of Logarithmic Scaling for Low Signals and Compensated Data
}
