% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fcs_files_df_get_ff.R
\name{fcs_files_df_get_ff}
\alias{fcs_files_df_get_ff}
\title{Convenience function to read many FCS files as flowframes into memory}
\usage{
fcs_files_df_get_ff(
  df,
  compensate = T,
  comp_keyword = "SPILL",
  match_channels = T
)
}
\arguments{
\item{df}{dafa frame with information to FCS files; as returned from fcexpr::wsp_get_gs or fcexpr::wsp_get_ff}

\item{compensate}{apply compensation with matrices from flowjo workspaces, obtained with fcexpr::wsx_get_compmat}

\item{comp_keyword}{the keyword under which the compensation is saved, usually SPILL or SPILLOVER;
passed to fcexpr::wsx_get_compmat}

\item{match_channels}{if channel names in compmat do not exactly match those in flowframes
(due to conversion of special characters), try to find the best matching channel name
and change it in the compmat; passed to ff_apply_compensation}
}
\value{
a list of flowframes
}
\description{
Read FCS files as flowframes into memory and have them compensated with compensation
matrices from the respective flowjo workspace. This function combines flowCore::read.FCS,
fcexpr::wsx_get_compmat and fcexpr::ff_apply_compensation. The flowframes my then be
subsetted with index matrices (ind_mat) as returned from fcexpr::wsp_get_ff or
fcexpr::wsp_get_indices.
}
\examples{
\dontrun{
# read subsetted flowframes (by population) via flowjo workspace
ffdata <- fcexpr::wsp_get_ff(wsp = "mypath/to/workspace.wsp", population = c("CD4", "CD8"), groups = "flowjogroup", FCS.file.folder = "local/path/to/fcsfiles")
# do the equivalent separately obtained comp matrix, indices (which rows belong to which population) and fcs files on disk
ffs <- fcexpr::fcs_files_df_get_ff(df = ffdata[["FCS_files"]])
# subset with ind mat
ff1 <- flowCore::Subset(ffs[[1]], ffdata[["ind_mats"]][[1]][,"/Time/Lymphocytes/Single Cells/CD3+/CD8"])
}
}
