% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fcs_get_keywords.R
\name{fcs_get_keywords}
\alias{fcs_get_keywords}
\title{Obtain keywords from FCS files}
\usage{
fcs_get_keywords(
  file_paths,
  keywords = NULL,
  return = c("data.frame", "vector")
)
}
\arguments{
\item{file_paths}{paths to fcs files}

\item{keywords}{selection of keywords to extract}

\item{return}{return format}
}
\value{
data frame of list of named vectors
}
\description{
Simple wrapper around flowCore::read.FCSheader with some error handling.
}
