% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fcs_get_voltages.R
\name{fcs_get_voltages}
\alias{fcs_get_voltages}
\title{Get voltages of PMTs from a fcs file}
\usage{
fcs_get_voltages(file_path)
}
\arguments{
\item{file_path}{character of path to the fcs file}
}
\value{
a data.frame with different columns depending on the machine the fcs was acquired with; usually the 'V'-column indicates PMT voltages
}
\description{
If there is a large number of fcs files and one wants to make sure that all files
have been acquired with the same settings this functions helps to pull out the
voltage of PMTs.
}
\examples{
\dontrun{
# When the script is saved to R_scripts in the experiment folder,
get the absolute path to the folder
wd <- dirname(dirname(rstudioapi::getActiveDocumentContext()$path))
# Find workspaces
ws <- list.files(wd, pattern = '\\\\.wsp$', recursive = T, full.names = T)
# Get groups from the first ws
gr <- lapply(ws[1], function(x) {
unique(as.character(CytoML::fj_ws_get_sample_groups(CytoML::open_flowjo_xml(x))$groupName))
})[[1]]
# pull out paths to fcs files of the first group
paths <- fcs_paths_from_ws(ws[1], gr[1])
# iterate (loop) through the paths to get the voltage of each fcs file
volts <- do.call(rbind, lapply(paths, function(x) {
v <- voltage_from_fcs(x)
v$file <- basename(x)
return(v)
}))
}
}
