% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fcs_write_comp_mat.R
\name{fcs_write_comp_mat_neutral}
\alias{fcs_write_comp_mat_neutral}
\title{Make an existing compensation matrix in a FCS neutral}
\usage{
fcs_write_comp_mat_neutral(
  fcs_file_path,
  comp_mat_keyword = "SPILL",
  channel_names = NULL
)
}
\arguments{
\item{fcs_file_path}{character, file path to the fcs file}

\item{comp_mat_keyword}{character, name of the keyword which holds the matrix}

\item{channel_names}{enter channels for comp matrix; only applied when matrix under comp_mat_keyword is missing of size 0}
}
\value{
no return, but FCS changed on disk
}
\description{
This will overwrite an existing compensation matrix in a FCS file so that
there are ones on the diagonal and else zeros. This is handy in order to
not having to type this manually in flow jo or so.
}
\examples{
\dontrun{
compMap_neutral_to_fcs(fcs_file_path = "my/path/to/FCSfile.fcs")
}
}
