% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ff_add_column.R
\name{ff_add_columns}
\alias{ff_add_columns}
\title{Add new columns to a flow frame}
\usage{
ff_add_columns(ff, mat = NULL, overwrite = F)
}
\arguments{
\item{ff}{flow frame}

\item{mat}{matrix with numeric columns}

\item{overwrite}{do overwrite existing columns in ff?}
}
\value{
flow frame
}
\description{
Add new columns to a flow frame
}
