% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ff_add_groups.R
\name{ff_add_groups}
\alias{ff_add_groups}
\title{Add sample group identifier to a flow frame}
\usage{
ff_add_groups(ff, ident_col = "ident", grouplist, overwrite = F)
}
\arguments{
\item{ff}{flow frame}

\item{ident_col}{column name with sample identifier}

\item{grouplist}{named list of sample information}

\item{overwrite}{do overwrite exisitng columns in ff}
}
\description{
Add sample group identifier to a flow frame
}
\examples{
\dontrun{
ff <- ff_merge(ff_list = list(ff1, ff2))
ff_add_groups(ff, grouplist = list(cond2 = c("d","k"), dis2 = c("e", "r")))
}
}
