% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ff_apply_compensation.R
\name{ff_apply_compensation}
\alias{ff_apply_compensation}
\title{Apply compensation matrix to flowframe}
\usage{
ff_apply_compensation(ff, compmat, match_channels = T)
}
\arguments{
\item{ff}{a flowframe}

\item{compmat}{a compensation matrix; matrix with column names}

\item{match_channels}{try to match channels names from ff with column names
of compmat}
}
\value{
compensated flowframe
}
\description{
In the end, it is flowCore::compensate only. But since channel names from flowframe
and column names of comp mats may not match, this function was made to handle that.
The problem arises due to conversion of special characters. Channel names may contain
a forward slash since this is how filters of a flow cytometer are described.
E.g. 760/30 for 760 nm +/- 15 nm. However, in FCS keywords the forward slash gets
converted to an underscore. Channel names in a flowframe maintain the slash. So,
flowCore::compensate complains.
}
\examples{
\dontrun{
compmats <- fcexpr::wsx_get_compmats(ws = path/to/flowjo/workspace.wsp)
ff <- flowCore::read.FCS("path/to/fcsfile.fcs", truncate_max_range = F)
correct_mat <- compmmats |> dplyr::fiilter(FileName == fcsfile.fcs) |> dplyr::pull(mat)
ff_compensated_exprs_slot <- ff_apply_compensation(ff, correct_mat)
}
}
