% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ff_calc_marker.R
\name{ff_calc_marker}
\alias{ff_calc_marker}
\title{Calculate cluster markers in data in a flow frame}
\usage{
ff_calc_marker(
  ff,
  exprs = NULL,
  channels = NULL,
  cluster_cols,
  mc.cores = 1,
  global = T,
  pairwise = F,
  ...
)
}
\arguments{
\item{ff}{flow frame}

\item{exprs}{expression matrix, if provided is preferred over ff}

\item{channels}{which channels to use, see ff_get_channels for selection
options}

\item{cluster_cols}{name of columns that contain cluster identities}

\item{mc.cores}{number of cores to use for calculation}

\item{global}{do calculate global cluster markers}

\item{pairwise}{do calculate pairwise cluster markers}

\item{...}{arguments to ff_get_channels}
}
\value{
list with data frames
}
\description{
Calculate cluster markers in data in a flow frame
}
