% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ff_calc_pca.R
\name{ff_calc_pca}
\alias{ff_calc_pca}
\title{Calculate PCA based on data in flow frame}
\usage{
ff_calc_pca(
  ff,
  channels = NULL,
  exprs = NULL,
  args = list(scale. = F, center = F),
  return = c("pca", "ff"),
  ...
)
}
\arguments{
\item{ff}{flow frame}

\item{channels}{which channels to use, see ff_get_channels for selection
options}

\item{exprs}{expression matrix, either provide ff or exprs}

\item{args}{arguments to stats::prcomp}

\item{return}{return pca object or flow frame with PC-channels added}

\item{...}{arguments to ff_get_channels}
}
\value{
flow frame or pca object
}
\description{
Calculate PCA based on data in flow frame
}
