% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ff_calc_som.R
\name{ff_calc_som}
\alias{ff_calc_som}
\title{Calculate self-organizing map based on flow frame}
\usage{
ff_calc_som(
  ff,
  channels = NULL,
  exprs = NULL,
  som_fun = EmbedSOM::SOM,
  som_args = list(),
  embedsom_args = list(),
  seed = 42,
  names = c("SOM_1", "SOM_2"),
  return = c("som", "ff"),
  ...
)
}
\arguments{
\item{ff}{flow frame}

\item{channels}{which channels to use, see ff_get_channels for selection
options}

\item{exprs}{expression matrix, either provide ff or exprs}

\item{som_fun}{EmbedSOM::SOM or EmbedSOM::GQTSOM}

\item{som_args}{arguments to som_fun}

\item{embedsom_args}{arguments to EmbedSOM::EmbedSOM}

\item{seed}{random seed}

\item{names}{names for SOM channels}

\item{return}{return SOM object or flow frame with SOM channels added}

\item{...}{arguments to ff_get_channels}
}
\value{
flow frame or SOM
}
\description{
Calculate self-organizing map based on flow frame
}
