% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ff_calc_umap.R
\name{ff_calc_umap}
\alias{ff_calc_umap}
\title{Calculate UMAP based on flow frame}
\usage{
ff_calc_umap(
  ff,
  channels = NULL,
  exprs = NULL,
  args = list(n_neighbors = 20, metric = "cosine", verbose = T, scale = T),
  seed = 42,
  names = c("UMAP_1", "UMAP_2"),
  return = c("umap", "ff"),
  ...
)
}
\arguments{
\item{ff}{flow frame}

\item{channels}{which channels to use, see ff_get_channels for selection
options}

\item{exprs}{expression matrix, either provide ff or exprs}

\item{args}{arguments to uwot::umap}

\item{seed}{random seed}

\item{names}{names of new dimensions to add as channels}

\item{return}{return UMAP object or flow frame with umap channels added}

\item{...}{arguments to ff_get_channels}
}
\value{
flow frame or umap object
}
\description{
Calculate UMAP based on flow frame
}
