% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ff_calc_umap_tsne.R
\name{ff_calc_umap_tsne}
\alias{ff_calc_umap_tsne}
\title{Calculate UMAP or tSNE dimension reduction based on flow frame}
\usage{
ff_calc_umap_tsne(
  ff,
  channels = NULL,
  exprs = NULL,
  args = list(n_neighbors = 20, metric = "cosine", verbose = T, scale = T),
  seed = 42,
  names = c("UMAP_1", "UMAP_2"),
  return = c("coords", "ff"),
  fun = uwot::umap,
  ...
)
}
\arguments{
\item{ff}{flow frame}

\item{channels}{which channels to use, see ff_get_channels for selection
options}

\item{exprs}{expression matrix, either provide ff or exprs}

\item{args}{arguments to fun}

\item{seed}{random seed}

\item{names}{names of new dimensions to add as channels}

\item{return}{return UMAP object or flow frame with umap channels added}

\item{fun}{uwot::umap or Rtsne::Rtsne}

\item{...}{arguments to ff_get_channels}
}
\value{
flow frame or umap object
}
\description{
Calculate UMAP or tSNE dimension reduction based on flow frame
}
