% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ff_compare_channels.R
\name{ff_compare_channels}
\alias{ff_compare_channels}
\title{Compare channel meta data of flow frames for equality}
\usage{
ff_compare_channels(ff_list, channels = NULL, strict = T)
}
\arguments{
\item{ff_list}{list of flow frames}

\item{channels}{relevant channels}

\item{strict}{accept differences like different desc}
}
\value{
NULL when there are no issues
}
\description{
Compare channel meta data of flow frames for equality
}
