% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ff_filter_obs.R
\name{ff_filter_obs}
\alias{ff_filter_obs}
\title{Filter observations from flowframes}
\usage{
ff_filter_obs(
  ff,
  channels = NULL,
  channels_select = F,
  limits = c(-Inf, Inf),
  limits_include = c(F, F),
  quantiles = c(0, 1),
  verbose = T
)
}
\arguments{
\item{ff}{flow frame or list of such}

\item{channels}{channels to use for filtering, goes into ff_get_channels}

\item{channels_select}{select channels from filtering? (remove other)}

\item{limits}{absolute limits for filtering}

\item{limits_include}{or >=}

\item{quantiles}{quantiles to use for filtering (relative by channel and ff)}
}
\value{
reduced ff
}
\description{
Remove extreme values that may disturb dimension reduction or plotting
}
\examples{
\dontrun{
fflist <- fcexpr::wsp_get_ff(w1,
                             FCS.file.folder = file.path(wd, "FCS_files"),
                             samples = grep("full", basename(f1[[1]]), value = T),
                             population = "Time")
ffs <- purrr::map(fflist[[1]], ~.x[[1]][[1]])
ffs <- ff_filter_obs(ffs, quantiles = c(0.001,0.999))
}
}
