% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ff_get_channels.R
\name{ff_get_channels}
\alias{ff_get_channels}
\title{Get channel names from flow frame}
\usage{
ff_get_channels(
  ff,
  timeChannel = NULL,
  channels = NULL,
  part_match = T,
  rm_scatter = T,
  rm_time = T,
  rm_wo_desc = F,
  replace_NA_desc = F
)
}
\arguments{
\item{ff}{flow frame}

\item{timeChannel}{time channel name}

\item{channels}{channels to return/match, can be channel name, desc or
keyword (see examples); if NULL all channels are returned}

\item{part_match}{allow partial matching of channels identifiers}

\item{rm_scatter}{remove scatter channels from return}

\item{rm_time}{remove time channel}

\item{rm_wo_desc}{remove all channels without desc (empty channels)}
}
\value{
character vector
}
\description{
Get channel names from flow frame
}
\examples{
\dontrun{
ff_get_channels(ff, channels = c("LP502", "PE"))
ff_get_channels(ff, channels = c("P4N", "P6N"))
ff_get_channels(ff, channels = c("CD3", "CD8"))
}
}
