% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ff_get_common_desc.R
\name{ff_get_common_desc}
\alias{ff_get_common_desc}
\title{Get intersecting channel description from flow frames}
\usage{
ff_get_common_desc(ff_list)
}
\arguments{
\item{ff_list}{list of flow frames}
}
\value{
character vector
}
\description{
Get intersecting channel description from flow frames
}
