% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ff_get_common_kw.R
\name{ff_get_common_kw}
\alias{ff_get_common_kw}
\title{Get intersecting keywords from flow frames}
\usage{
ff_get_common_kw(ff_list)
}
\arguments{
\item{ff_list}{list of flow frames}
}
\value{
list of keywords
}
\description{
Get intersecting keywords from flow frames
}
