% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ff_get_exprsdf.R
\name{ff_get_exprsdf}
\alias{ff_get_exprsdf}
\title{Turn subset of ff exprs slot into data frame}
\usage{
ff_get_exprsdf(ff, channels = c("umap", "som", "res\\\\.", "ident"), ...)
}
\arguments{
\item{ff}{flow frame}

\item{channels}{channels to include, passed to ff_get_channels}

\item{...}{arguments to ff_get_channels}
}
\value{
data frame
}
\description{
If there columns like ident or other grouplist they are recoded back
from numeric to character. attributes(ff) has to contain grouplist.
}
