% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ff_merge.R
\name{ff_merge}
\alias{ff_merge}
\title{Combine/merge flow frames}
\usage{
ff_merge(
  ff_list,
  ident_col = "ident",
  grouplist = NULL,
  keywords_to_channels = NULL,
  add_transformed_channels = F,
  trafoname = "trafolistinv",
  suffix = "_trans",
  suffix_trafo = "",
  exclude = fcexpr:::non_fluo_channels,
  exclude_from = c("raw", "trafo"),
  order_first = c("trafo", "raw")
)
}
\arguments{
\item{ff_list}{list of flow frames}

\item{ident_col}{identity column to add}

\item{grouplist}{other sample info to add, see ff_add_groups}

\item{keywords_to_channels}{keywords to add as channels; $-signs are
removed}

\item{add_transformed_channels}{also add transformed channels to final ff;
all arguments below only become relevant when this is TRUE}

\item{trafoname}{transformation name to use , must be in attributes of ff}

\item{suffix}{suffix to add to channel names without trafo}

\item{suffix_trafo}{suffix to add to channel names that were transformed}

\item{exclude}{channels to exclude from transformation}

\item{exclude_from}{were to exclude those channels from, the transformed
or not-transformed channels}

\item{order_first}{which channels to appear first in ff}
}
\value{
flow frame
}
\description{
Default arguments assume that ff in ff_list were not inverse transformed
during import.
}
