% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ff_plot_quantiles.R
\name{ff_plot_quantiles}
\alias{ff_plot_quantiles}
\title{Plot quantiles and their values channel-wise and flowframe-wise}
\usage{
ff_plot_quantiles(ff, channels = NULL, quantiles = seq(0, 1, 0.01))
}
\arguments{
\item{ff}{one flowframe or a list of such}

\item{channels}{channels to plot, goes into ff_get_channels}

\item{quantiles}{which quantils to calculate}
}
\value{
list
}
\description{
Plot quantiles and their values channel-wise and flowframe-wise
}
