% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ff_rename_channels.R
\name{ff_rename_channels}
\alias{ff_rename_channels}
\title{Rename one or many channels of a flowframe}
\usage{
ff_rename_channels(ff, rename_vec)
}
\arguments{
\item{ff}{flowframe}

\item{rename_vec}{named vector with current channel name as name
and new names as values (see example)}
}
\value{
updated flowframe
}
\description{
Alter all slots: exprs, params, desc (keys).
}
\examples{
\dontrun{
ff <- flowCore::read.FCS("path/to/file/Specimen_001_Tube_001_001.fcs")
colnames(flowCore::exprs(ff)) # inspect
rename_vec <- setNames(c("Pe", "APC"), c("yg-586/15-E-A", "r-670/30-C-A"))
ffnew <- ff_rename_channels(ff, rename_vec)
# optionally: overwrite fcs file on disk
flowCore::write.FCS(ffnew, filename = "path/to/file/Specimen_001_Tube_001_001.fcs")
}
}
