% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ff_to_mat.R
\name{ff_to_mat}
\alias{ff_to_mat}
\title{Get expression matrix from a flow frame}
\usage{
ff_to_mat(
  ff,
  grep = c("trans", "umap", "som", "tsne", "ident"),
  channels = NULL
)
}
\arguments{
\item{ff}{flowframe}

\item{grep}{character to select channel}

\item{channels}{in addition to grep, pass channels to ff_get_channels}
}
\value{
matrix or data frame
}
\description{
Channel descriptions from parameters are used as colnames.
}
\examples{
\dontrun{
mat <- ff_to_mat(ff)
}
}
