% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ff_transform.R
\name{ff_transform}
\alias{ff_transform}
\title{Transform a flow frame}
\usage{
ff_transform(ff, trafolist)
}
\arguments{
\item{ff}{flow frame}

\item{trafolist}{matching transformation list}
}
\value{
flow frame
}
\description{
Intended to switch between inverse.transform = T or F as returned by
flowWorkspace::gh_pop_get_data.
}
\examples{
\dontrun{
fflist <- fcexpr::wsp_get_ff(ws)
ff <- fflist[["flowframes"]][[1]][[1]][["transformed"]]
fflist[["flowframes"]][[1]][[1]][["untransformed"]] <- ff_transform(ff = ff, trafollist = attr(ff, "trafolistinv"))
ff_un <- fflist[["flowframes"]][[1]][[1]][["untransformed"]]
fflist[["flowframes"]][[1]][[1]][["transformed2"]] <- ff_transform(ff = ff_un, trafollist = attr(ff, "trafolist"))
}
}
