% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ff_update.R
\name{ff_update}
\alias{ff_update}
\title{Update params and keywords when exprs was changed}
\usage{
ff_update(ff)
}
\arguments{
\item{ff}{flowframe}
}
\value{
flowframe
}
\description{
Update params and keywords when exprs was changed
}
\examples{
\dontrun{
ff <- ff_update(ff)
}
}
