% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geo_mean.R
\name{geo_mean}
\alias{geo_mean}
\title{Geometric mean}
\usage{
geo_mean(x, rm.na = F)
}
\arguments{
\item{x}{numeric vector}

\item{rm.na}{logical if to remove na or not}
}
\value{
a numeric representing the geomtric mean of values in x
}
\description{
In flow cytometry the geometric mean is a common metric of central tendency.
It is less susceptible to outlier than the arithmetic mean, but more susceptible
as compared to the median. If you have untransformed values of fluorescence
intensities (original output of the machines) this function may not work as you
may have negative values in your vector x. Hence, the median has to be used or
the data needs transformation (e.g. logicle transformation).
}
\examples{
geo_mean(c(1,2,3,10,100))
}
