% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ws_get_popstats.R
\name{get.popStats.from.ws}
\alias{get.popStats.from.ws}
\title{Title}
\usage{
get.popStats.from.ws(ws, gr, FCS.file.folder, groupwise = T)
}
\arguments{
\item{ws}{character}

\item{gr}{character}

\item{FCS.file.folder}{character}

\item{groupwise}{logical}
}
\value{
data.frame
}
\description{
Title
}
\examples{
\dontrun{
ws_get_popstats(ws = "fj.wsp", gr = "full.stain", FCS.file.folder = "FCS.files", groupwise = T)
}
}
