% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_hclust_clusters.R
\name{get_hclust_clusters}
\alias{get_hclust_clusters}
\title{Calculate clusters with hclust}
\usage{
get_hclust_clusters(
  exprs,
  distance = c("euclidean", "correlation", "maximum", "manhattan", "canberra", "binary",
    "minkowski"),
  method = c("complete", "ward.D", "ward.D2", "ward", "single", "average", "mcquitty",
    "median", "centroid"),
  k = c(5, 10, 20),
  mc.cores = 1
)
}
\arguments{
\item{exprs}{numeric matrix with channels as columns}

\item{distance}{distance metric for stats::dist}

\item{method}{method for stats::hclust}

\item{k}{number of clusters to generate by stats::cutree}
}
\value{
matrix with cluster annotations as columns
}
\description{
stats::dist takes quite some time, reduction of columns with pca may help
}
