% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gs_get_descname_lookup.R
\name{gs_get_descname_lookup}
\alias{gs_get_descname_lookup}
\title{Channel names / description lookup vector}
\usage{
gs_get_descname_lookup(gs)
}
\arguments{
\item{gs}{a gatingset}
}
\value{
a named vector
}
\description{
All gating hierarchies (gh) in gs are checked for equal channel descriptions.
For channel names with unique desc across all gh, desc are returned,
otherwise the channel name.
}
\examples{
\dontrun{
# read gatingsets
gslist1 <- wsp_get_gs(wsp1)
lookup <- gs_get_descname_lookup(gslist1[[1]][[1]])
}
}
