% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gslist_try_merge.R
\name{gslist_try_merge}
\alias{gslist_try_merge}
\title{Try to merge gatingsets by removing non-common populations/gates}
\usage{
gslist_try_merge(gs_list)
}
\arguments{
\item{gs_list}{list of gatingsets}
}
\value{
gatingset
}
\description{
Uses flowWorkspace::merge_list_to_gs
}
