% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/guess_marker_channels.R
\name{guess_marker_channels}
\alias{guess_marker_channels}
\title{Narrow channels of stained antibody}
\usage{
guess_marker_channels(popstats_df)
}
\arguments{
\item{popstats_df}{data frame from wsx_get_popstats}
}
\value{
data frame
}
\description{
Narrow channels of stained antibody
}
\examples{
\dontrun{
df <- wsx_get_popstats(ws_path, strip_data = F)[[1]]
df2 <- guess_marker_channels(df)
}
}
