% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatmap_ordering.R
\name{heatmap_ordering}
\alias{heatmap_ordering}
\title{Order columns and rows of a long data frame for an appealing heatmap}
\usage{
heatmap_ordering(
  df,
  features = "channel",
  groups = "cluster",
  values = "mean_cluster_scale",
  feature_order = c("custom", "hclust", "none"),
  group_order = c("hclust", "none", "custom")
)
}
\arguments{
\item{df}{data frame in long format}

\item{features}{column name with features}

\item{groups}{column name with groups}

\item{values}{column name that becomes values in heatmap}

\item{feature_order}{how to order features}

\item{group_order}{how to order groups}
}
\value{
data frame with updated factor levels for rows and cols
}
\description{
Heatmaps only become meaningful when rows and columns are ordered properly.
If features or groups column of df are factors, then this remains unchanged
if order is 'none'. That way either order can be fixed as desired.
If both methods are set to 'hclust' this corresponds
to the procedure in pheatmap::pheatmap.'Custom' is a another heurisitc to
order features, preferentially. The default settings provide a good
basis, you may pre-set an order for columns by defining factors and leave
row_order 'custom'.
}
