% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inds_get_ff.R
\name{inds_get_ff}
\alias{inds_get_ff}
\title{Title}
\usage{
inds_get_ff(
  ind_mat,
  population,
  alias_attr_name = "short_names",
  path_attr_name = "FilePath",
  downsample = 1,
  inverse_transform = c(T, F),
  lapply_fun = lapply,
  ...
)
}
\arguments{
\item{...}{}
}
\value{

}
\description{
Title
}
