% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/int_breaks.R
\name{int_breaks}
\alias{int_breaks}
\title{Make axis labels that are numeric but should only integer values}
\usage{
int_breaks(n = 5, ...)
}
\arguments{
\item{n}{integer giving the desired number of intervals. Non-integer values are rounded down. (passed to base::pretty)}

\item{...}{further arguments passed to base::pretty}
}
\value{
a vector or axis labels of length n
}
\description{
Make axis labels that are numeric but should only integer values
}
\examples{
int_breaks(n=3)(c(1,4,5,6,7,10))
int_breaks(n=6)(c(1.1,4.9,5.8,6,7,10.78))
# or pass to ggplot2: ggplot2::scale_x_continuous(breaks = int_breaks(n = 4))
}
