% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_hd_pops.R
\name{make_hd_pops}
\alias{make_hd_pops}
\title{Create a high dimensional space with multiple populations}
\usage{
make_hd_pops(npop = 5, popindexstart = 1, rm_ann_cols = F, ...)
}
\arguments{
\item{npop}{number of population to generate}

\item{popindexstart}{identity of first population}

\item{rm_ann_cols}{remove annotation columns like cell and pop;
keep protein cols only}

\item{...}{arguments to create_pop}
}
\value{
data frame
}
\description{
Create a high dimensional space with multiple populations
}
\examples{
# default settings
df1 <- make_hd_pops()
# plot with ggplot
ggplot(df1, aes(protein1, protein2)) +
  geom_point(aes(color = pop)) +
  scale_x_log10() +
  scale_y_log10() +
  brathering::scale_color_custom()
ggplot(df1, aes(protein3, protein4)) +
  geom_point(aes(color = pop)) +
  scale_x_log10() +
  scale_y_log10() +
  brathering::scale_color_custom()
# dimension reduction with umap
df11 <- cbind(df1, as.data.frame(uwot::umap(as.matrix(df1[,1:5]), verbose = T)))
brathering::plot2(df11[,c("V1", "V2", "pop", "group")], cex = 3, color = "pop", legend = NULL)
brathering::plot3d(df1, transform = "log10")
#add pops with other properties
df2 <- dplyr::bind_rows(df1, make_hd_pops(popindexstart = 6,
                                          range_sd_fct = 2,
                                          range_modes = 1))
ggplot(df2, aes(protein1, protein2)) +
  geom_point(aes(color = pop)) +
  scale_x_log10() +
  scale_y_log10() +
  brathering::scale_color_custom()
}
