% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/match_fcs_and_wsp.R
\name{match_fcs_and_wsp}
\alias{match_fcs_and_wsp}
\title{Find out in which FlowJo workspaces FCS files have been analyzed.}
\usage{
match_fcs_and_wsp(
  FCS.file.folder = NULL,
  FCS.file.paths = NULL,
  exclude.folders = NULL,
  wsp
)
}
\arguments{
\item{FCS.file.folder}{one or more absolute paths to folder containing FCS files;
will be scanned for .fcs files recursively (subfolder as well)}

\item{FCS.file.paths}{absolute paths to to FCS files}

\item{exclude.folders}{folders to exclude; if any of these strings appears
in the absolute path to a fcs file this file will be excluded}

\item{wsp}{one or more absolute paths to flowjo workspaces on disk}
}
\value{
a data.frame of matched (joined) information about fcs files on disk
and fcs files referred to in flowjo workspaced
}
\description{
Find out in which FlowJo workspaces FCS files have been analyzed.
}
\examples{
\dontrun{
match_fcs_and_wsp(FCS.file.folder = file.path(wd, "FCS_files"), ws = "mypath/my.wsp")
}
}
