% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_exp.R
\name{new_exp}
\alias{new_exp}
\title{Create an experiment folder for flow cytometry experiments}
\usage{
new_exp(path = NULL, name = NULL, date_prefix = T)
}
\arguments{
\item{path}{path to the parent directory where the experiment folder is to be initiated}

\item{name}{name of the experiment folder}

\item{date_prefix}{logical if the folder name should be prefixed with the current date}
}
\value{
No return value. Instead a folder with template files is written to disk.
}
\description{
Create an experiment folder for flow cytometry experiments
}
\examples{
\dontrun{
new_exp(path = '/Users/CMS/Documents/experiments', name = 'CD3_titration')
}
}
