% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_hd_pops.R
\name{prob_vec}
\alias{prob_vec}
\title{Create a probability vector}
\usage{
prob_vec(
  n = 1000,
  type = c("U", "M"),
  power = 5,
  peakshift = c(0.2, 0.2),
  beta_ab = c(0.2, 0.4)
)
}
\arguments{
\item{n}{number of elements}

\item{type}{type, roughly shape}

\item{power}{for M-type, exponent to initial probs, higher power --> higher peaks}

\item{peakshift}{for M-type, peakshift towards middle}

\item{beta_ab_range}{for U-type, range of ab values to dbeta, can also
be a vector of one or two values only}
}
\value{
numeric vector which sums to 1
}
\description{
Create a probability vector
}
\examples{
library(ggplot2)
# type U: low beta_ab_range --> steep ends, high beta_ab_range --> close to unif
probs <- purrr::map_dfr(setNames(c(0.3,0.5,0.75,0.95),c(0.3,0.5,0.75,0.95)),
                        ~data.frame(x = prob_vec(type = "U", beta_ab = .x)) |>
                          dplyr::mutate(index = dplyr::row_number()),
                        .id = "beta_ab")
ggplot(probs, aes(index,x)) +
  geom_line() +
  facet_wrap(vars(beta_ab))
# type M: higher power to increase peaks
probs <- purrr::map_dfr(setNames(c(2,4,8,12),c(2,4,8,12)),
                        ~data.frame(x = prob_vec(type = "M", power = .x)) |>
                          dplyr::mutate(index = dplyr::row_number()),
                        .id = "power")
ggplot(probs, aes(index,x)) +
  geom_line() +
  facet_wrap(vars(power))
}
