% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_fcs.R
\name{read_fcs}
\alias{read_fcs}
\title{Read FCS file with flowCore and non_modifying defaults}
\usage{
read_fcs(file_path, truncate_max_range = F, transformation = F, ...)
}
\arguments{
\item{file_path}{path to to fcs file}

\item{truncate_max_range}{see flowCore::read.FCS}

\item{transformation}{see flowCore::read.FCS}

\item{...}{args to flowCore::read.FCS}
}
\description{
This simplest wrapper to imagine just to pass default arguments which
avoid any data manipulation
}
