% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampledescription_to_fcs_keywords.R
\name{sampledescription_to_fcs_keywords}
\alias{sampledescription_to_fcs_keywords}
\title{Write columns from a sampledescription as keyword into FCS files}
\usage{
sampledescription_to_fcs_keywords(sampledescription, columns, FCS.file.folder)
}
\arguments{
\item{sampledescription}{data frame}

\item{columns}{vector of column names from sampledescription to write as keywords}

\item{FCS.file.folder}{folder of FCS files}
}
\value{
no return but keywords written to FCS files
}
\description{
ATTENTION: This function will alter the meta data of FCS files
such that one can notice it was once opened and saved with flowCore.
}
