% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sort_num_chr.R
\name{sort_num_chr}
\alias{sort_num_chr}
\title{Sort numerical character according to the numbers that the character values represent.}
\usage{
sort_num_chr(x, decreasing = F)
}
\arguments{
\item{x}{vector of character or factor which represent numbers}

\item{decreasing}{logical, sort decreasing}
}
\value{
character vector of sorted values
}
\description{
Just a convenience function often used to sort patient IDs by numerical value
while being considered as factor though.
}
\examples{
sort_num_chr(c("3", "4", "2"))
sort_num_chr(as.factor(c("3", "4", "2")))
}
