% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split_flowjo_export_image.R
\name{split_flowjo_export_image}
\alias{split_flowjo_export_image}
\title{Split a flowjo export of multiple plots into separate graphics for rearrangement}
\usage{
split_flowjo_export_image(
  img,
  nPlots_x = 4,
  nPlots_y = 3,
  folder,
  pptx_name,
  pptx_image_size = 1,
  pptx_border_space = 0.2
)
}
\arguments{
\item{img}{character path to an image, preferable a png exported from flowjos layout editor, has to be equally spaced in x and y direction}

\item{nPlots_x}{integer, how many plots in x direction}

\item{nPlots_y}{integer, how many plots in y direction}

\item{folder}{folder name of where to save split images, this folder is written to the directory of img}

\item{pptx_name}{character, name a pptx to write the split images to; if missing no pptx is generated}

\item{pptx_image_size}{numeric, size of the images in pptx}

\item{pptx_border_space}{numeric, space between images in pptx}
}
\value{
writes a folder with split images and optionally a pptx
}
\description{
Split a flowjo export of multiple plots into separate graphics for rearrangement
}
\examples{
\dontrun{
# copy an example file to your disk
file.copy(from = system.file("extdata", "img.png", package = "fcexpr"), to = "your_path")
fcexpr::split_flowjo_export_image(img = "your_path/img.png",
nPlots_x = 6, nPlots_y = 6, pptx = "my_img_split.pptx")
}
}
