% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ws_get_MFIs.R
\name{ws_get_MFIs}
\alias{ws_get_MFIs}
\title{Retrieve MFIs from populations gated in one or more flowjo workspaces}
\usage{
ws_get_MFIs(
  ws,
  gr,
  FCS.file.folder,
  mean.fun = "median",
  variance.fun = "sd",
  inverse.transform = T
)
}
\arguments{
\item{ws}{character vector of paths to flowjo workspaces}

\item{gr}{character vector of flowjo groups to import or a list of those vectors, each for one ws}

\item{FCS.file.folder}{path to root folder which contains FCS files; if missing file.path(getwd(), "FCS_files") is assumed}

\item{mean.fun}{character name of the function name to use to calculate the average FI}

\item{variance.fun}{character name of the function name to use to calculate the variacnce of FI}

\item{inverse.transform}{logical indicating if fluorescence values are to be inversely transformed}
}
\value{
data.frame
}
\description{
If inverse.transform is set to TRUE values as shown by flowjo will be used
for calculation. This may omit to use 'geo_mean' for mean.fun as negative
values are not allowed. Inverse.transform = F leads to data transformation by
the logicle transform.
}
\examples{
\dontrun{
# When the script is saved to R_scripts in the experiment folder,
# get the absolute path to the folder
wd <- dirname(dirname(rstudioapi::getActiveDocumentContext()$path))
# find workspaces
ws <- list.files(wd, pattern = "\\\\.wsp$", recursive = T, full.names = T)
# get groups
gr <- lapply(ws, function(x) {
unique(as.character(CytoML::fj_ws_get_sample_groups(CytoML::open_flowjo_xml(x))$groupName))
})
# read the compensated values from flowjo and calculate MFIs
ws_get_MFIs(ws = ws, gr = gr)
}
}
