% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wsp_get_ff.R
\name{wsp_get_ff}
\alias{wsp_get_ff}
\title{Get (subsetted) flowFrames from one or many flowjo workspaces}
\usage{
wsp_get_ff(
  wsp,
  FCS.file.folder = NULL,
  groups = NULL,
  population = "root",
  invert_groups = F,
  samples = NULL,
  invert_samples = F,
  return_untransformed = F,
  return_transformed = T,
  downsample = 1,
  remove_redundant_channels = F,
  lapply_fun = lapply,
  seed = 42,
  downsample_channels = NULL,
  downsample_leverage_score = F,
  return_ind_mat = F,
  dir = tempdir(),
  channels = NULL,
  ...
)
}
\arguments{
\item{wsp}{vector of paths to flowjo workspaces}

\item{FCS.file.folder}{path to folder(s) of FCS files; may be one path for all wsp or a vector of paths, one for each wsp;
if not provided fcs file paths are derived individually from the wsp (xml). If the workspace was generated and saved on
another computer you will have to provide the path to FCS files on the current computer.}

\item{groups}{vector or list of groups in flowjo to consider; if a list, each index corresponds to the same index in wsp;
if NULL samples from all groups are read}

\item{population}{which population (=node, =gate) to subset flowFrames on; use fcexpr::wsx_get_poppaths to get paths}

\item{invert_groups}{logical whether to invert group selection}

\item{samples}{vector or list of samples to select (names of FCS files), each index corresponds to the index in wsp;
if NULL all samples (from selected groups) are read}

\item{invert_samples}{logical whether to invert sample selection}

\item{return_untransformed}{logical; do return untransformed (inverse) data}

\item{return_transformed}{logical; do return transformed data (transformation as in flowjo?!)}

\item{downsample}{numeric, if < 0 then a fraction of events is sampled, if > 0 an absolute number of events is sampled; or set to "min"
which will lead to downsampling each flowframe to the number of events in the flowframe with lowest number of events; can be a single value to treat all
value for all FCS files or a vector of same length as FCS files}

\item{remove_redundant_channels}{remove channels that are not part of the gating tree, mainly to reduce memory load}

\item{lapply_fun}{lapply function name, unquoted; lapply, pbapply::pblapply or parallel::mclapply are suggested}

\item{seed}{set a seed to reproduce downsampling}

\item{downsample_channels}{channels to use for leverage score calculation; use wsx_get_keywords to retrieve channel names/descriptions}

\item{downsample_leverage_score}{logical whether to use leverage scores for downsampling}

\item{return_ind_mat}{return full index matrix?}

\item{dir}{direction to save .h5 files to}

\item{channels}{channel names to filter flow frames for; when named vector, names are added as channel desc}

\item{...}{additional argument to the lapply function; mainly mc.cores when parallel::mclapply is chosen}
}
\value{
a list of (subsetted) flowframes with events that are within the gated population only
}
\description{
From a flowjo workspace with gated populations and the respective fcs files flowframes are generated.
Under the hood CytoML::flowjo_to_gatingset applies the geometric gate definitions and filters
relevant rows (indices) of fcs files. The compensation matrix as defined in flowjo will be used
to compensate fluorescence intensities. The index matrix (ind_mat) returned cannot be used with
the flow frames as these contain subsetted rows. Rather ind_mat can be used to read defined
rows from the fcs files without the need to re-read gate definitions from the workspace again.
ind_mats only, without any flow frames of gated populations, are returned by fcexpr::wsp_get_indices.
When reading data with ind_mats from fcs files, compensation needs to be applied manually. One
has to extract the compensation matrix either from the fcs file or from the flowjo workspace
where it is usually defined.
}
\examples{
\dontrun{
ff_list <- fcexpr::wsp_get_ff(wsp = "mypath/my.wsp", population = c("CD8+", "CD4+"), groups = "Compensation", invert_groups = TRUE)
# ffs <- purrr::map(ff_list[[1]], ~.x[[1]][[1]])
# ffs may be passed to fcexpr::dimred_to_fcs
# consider fcexpr::ff_filter_obs before that
## how to to work with the returned nested list
# ignore ind_mats and create one big data.frame w/o any filtering from flow frames
ff1 <- purrr::list_flatten(x = purrr::list_flatten(x = ff.list[["flowframes"]],
                                                   name_spec = "{outer}___{inner}"),
                           name_spec = "{outer}___{inner}")
ff1 <-
  purrr::map_dfr(ff1, purrr::compose(tibble::as_tibble, flowCore::exprs), .id = "type") |>
  tidyr::separate(type, into = c("sample", "pop", "trafo"), sep = "___")

# pluck a subset of flowframe only
ff2 <- purrr::map(ff.list[["flowframes"]], purrr::pluck, "CD4", "transformed")
ff2 <- purrr::map_dfr(ff2, purrr::compose(tibble::as_tibble, flowCore::exprs), .id = "sample")

# discard a transformation, then extract expression values
ff3 <- purrr::map(ff.list[["flowframes"]], ~ purrr::map(.x, ~ purrr::discard(.x, names(.x) == "untransformed")))
ff3 <- purrr::map(ff3, ~ purrr::map(.x, ~ purrr::pluck(.x, "transformed")))
ff3 <- purrr::map(ff3, ~ purrr::map(.x, ~ tibble::as_tibble(flowCore::exprs(.x))))

## how to work with ind_mats
# see ?fcexpr::inds_get_ff
# or ?fcexpr::fcs_files_df_get_ff
}
}
