% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wsx_compMats_to_fcs.R
\name{wsx_compMats_to_fcs}
\alias{wsx_compMats_to_fcs}
\title{Write the compensation matrices generated in a flowjo workspace to the SPILL keywords of respective FCS files}
\usage{
wsx_compMats_to_fcs(ws, groups = NULL, alt_FCS_file_folder = NULL, ...)
}
\arguments{
\item{ws}{path to flowjo workspace}

\item{groups}{which flowjo groups to get fcs files from}

\item{alt_FCS_file_folder}{if FCS files paths on disk are not the same as
in flowjos wsp file, then provide the correct path here}

\item{...}{additional arguments to fcexpr:::prep_spill()}
}
\value{
no return but SPILL keyword updated in FCS files
}
\description{
The compensation matrices generated or optimized in FlowJo only exist in the workspace. The underlying FCS files remain unchanged.
This is part of FlowJos philosophy, FCS remain original all the time. This is actually not a bad procedure but reduces flexibility in some cases.
This function saves the associated compensation matrices to the FCS files - the matrices are written into the SPILL keyword. That makes
them easily and unambiguously available for other calculations outside of FlowJo (e.g. in R) when compensated channels are required.
}
\examples{
\dontrun{
wsx_compMats_to_fcs(ws = "mypath/my.wsp")
}
}
