% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wsx_get_compmat.R
\name{wsx_get_compmat}
\alias{wsx_get_compmat}
\title{Get compensation matrices from flowjo workspace}
\usage{
wsx_get_compmat(ws, comp_keyword = "SPILL")
}
\arguments{
\item{ws}{path to flowjo workspace}

\item{comp_keyword}{SPILL or SPILLOVER}
}
\value{
data frame with FileNames and prepared matrices
}
\description{
From xml keywords the compensation matrices are obtained. Keyword is SPILL, usually.
The matrix first comes as one string which is converted to a matrix with
fcexpr:::compmat_key_to_mat. Matrices created in flowjo only exist in the xml (wsp) file.
To write them to fcs files, use fcexpr::compMat_to_FCS.
}
\examples{
\dontrun{
compmat_df <- fcexpr::wsx_get_compmat(ws = "path/to/flowjo/workspace.wsp")
}
}
