% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wsx_get_fcs_paths.R
\name{wsx_get_fcs_paths}
\alias{wsx_get_fcs_paths}
\title{Get file paths to the fcs files in a workspace}
\usage{
wsx_get_fcs_paths(ws, split = T, filter_AllSamples = F)
}
\arguments{
\item{ws}{path to flowjo workspace or a parsed xml-document (xml2::read_xml(ws))}

\item{split}{logical whether to split file paths by groups; if TRUE a list is returned, if FALSE a data frame}

\item{filter_AllSamples}{logical whether to filter the All Samples Group in case the fcs file is also part of another group}
}
\value{
a list or data frame of paths to fcs files
}
\description{
If one has a large experiment with many fcs files and many workspaces with various
gatings, finding out which fcs files are gated in which workspace may be helpful.
}
\examples{
\dontrun{
wsx_get_fcs_paths(ws)
}
}
