% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wsx_get_gate_definitions.R
\name{wsx_get_gate_definitions}
\alias{wsx_get_gate_definitions}
\title{Obtain geometric gate definitions (coordinates) from flowjo workspace}
\usage{
wsx_get_gate_definitions(ws, groups = NULL, invert_groups = F)
}
\description{
Obtain geometric gate definitions (coordinates) from flowjo workspace
}
