% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wsx_get_groups.R
\name{wsx_get_groups}
\alias{wsx_get_groups}
\title{Retrieve groups within a flowjo workspace and associated samples (sampleID)}
\usage{
wsx_get_groups(
  ws,
  filter_AllSamples = T,
  collapse = "nest",
  force_collapse = F
)
}
\arguments{
\item{ws}{path to flowjo workspace or a parsed xml-document (xml2::read_xml(ws))}

\item{filter_AllSamples}{logical whether to filter the All Samples Group in case the fcs file is also part of another group}

\item{collapse}{string how to collapse groups; to collapse to a list-column pass 'list'; to collapse to a string provide any separator string like ';', ',' or '\emph{-}'}

\item{force_collapse}{collapse in any case, even though there is only one group per sample}
}
\value{
a data frame
}
\description{
Retrieve groups within a flowjo workspace and associated samples (sampleID)
}
\examples{
\dontrun{
wsx_get_groups(ws)
}
}
