% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wsx_get_poppaths.R
\name{wsx_get_poppaths}
\alias{wsx_get_poppaths}
\title{Obtain paths to populations/nodes/gates in a flowjo workspace}
\usage{
wsx_get_poppaths(ws, groups = NULL, collapse = T)
}
\arguments{
\item{ws}{path to flowjo workspace or a parsed xml-document (xml2::read_xml(ws))}

\item{groups}{which groups in from flowjo workspace to consider}

\item{collapse}{logical whether to collapse FileNames to a list column}
}
\value{
data frame
}
\description{
Obtain paths to populations/nodes/gates in a flowjo workspace
}
\examples{
\dontrun{
pp <- wsx_get_poppaths(ws, collapse = F)
# check which files have equal gating trees
# and get the node (or path, or population) names
pp <- pp |>
dplyr::group_by(PopulationFullPath, Population, ws) |>
dplyr::summarise(FileName = list(FileName), .groups = "drop")
}
}
