% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wsx_get_popstats_0.R
\name{wsx_get_popstats_0}
\alias{wsx_get_popstats_0}
\title{Import data from a wsp file}
\usage{
wsx_get_popstats_0(
  ws,
  groups = NULL,
  invert_groups = F,
  return_stats = F,
  strip_data = T
)
}
\arguments{
\item{ws}{path to flowjo workspace or a parsed xml-document (xml2::read_xml(ws))}

\item{groups}{vector of flowjo group names to consider}

\item{invert_groups}{logical whether to exclude the selected groups}

\item{return_stats}{logical (T,F) whether to return statistics next to cells counts}

\item{strip_data}{remove internal columns from count df and speed up the function}
}
\value{
list with counts and statistics and graphs
}
\description{
Flowjo wsp files contain many information like gated event counts, statistics and keywords
from FCS files. These may be accessed without a dongle and can be read completely independent of
the respective FCS files once the gating has been conducted.
}
\examples{
\dontrun{
# When the script is saved to R_scripts in the experiment folder,
# get the absolute path to the folder
wd <- dirname(dirname(rstudioapi::getActiveDocumentContext()$path))
# find workspaces
ws <- list.files(path = wd, pattern = '\\\\.wsp$', recursive = T, full.names = T)
# import the population counts:
lst <- wsx_get_popstats(ws = ws[[1]])
#' # plot graph
ggraph::ggraph(lst$graph_sample[[1]]) +
  ggraph::geom_node_point() +
  ggraph::geom_edge_link() +
  ggraph::geom_node_label(ggplot2::aes(label = name))
}
}
