% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wsx_get_sample_df.R
\name{wsx_get_sample_df}
\alias{wsx_get_sample_df}
\title{Get information on samples as in ws and local FCS files that match those in ws}
\usage{
wsx_get_sample_df(
  ws,
  groups = NULL,
  invert_groups = F,
  samples = NULL,
  invert_samples = F,
  FCS.file.folder
)
}
\arguments{
\item{ws}{vector of paths to flowjo workspace}

\item{groups}{flowjo groups to consider, one vector for all ws or a list with one vector for each ws}

\item{invert_groups}{consider all other flowjo groups found in ws but groups}

\item{samples}{samples (FCS filenames) to consider, one vector for all ws or a list with one vector for each ws}

\item{invert_samples}{consider all other samples found in ws but samples}

\item{FCS.file.folder}{one path to the top local FCS file folder or a vector}
}
\value{
data frame with sample data
}
\description{
This function is used within wsp_get_ff, wsp_get_gs and wsp_get_indices. Usage
by the user may be followed by fcexpr::fcs_files_df_get_ff to import the FCS files
as flow frames easily. With ind_mats from fcexpr::wsp_get_indices the flowframes
can be subsetted by gated populations. See wsp_get_ff, wsp_get_gs and wsp_get_indices
for explanation of parameters (function arguments).
}
