% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wsx_get_sample_df_light.R
\name{wsx_get_sample_df_light}
\alias{wsx_get_sample_df_light}
\title{Get information on samples in a flowjo workspace}
\usage{
wsx_get_sample_df_light(ws)
}
\arguments{
\item{ws}{path to a flowjo workspace}
}
\value{
list of data frames
}
\description{
Info on group belonging, some keywords which form the identity of FCS files
and location on disk as stored in the ws. Th summary is a collapsed version
which can be used to check whether there are sample in one FlowJoGroup that
have the same $FIL keyword(=name). Because CytoML::flowjo_to_gatingset may
complain that one of these files does not match, if either one is to be
imported. In comparison to wsx_get_sample_df, local FCS file existence (as on
disk on the current machine is not checked); also no sample selection takes
place here.
}
