% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wsx_get_transformations.R
\name{wsx_get_transformations}
\alias{wsx_get_transformations}
\title{Get transformation parameters from a flowjo workspace}
\usage{
wsx_get_transformations(
  ws,
  lapply_fun = lapply,
  groups = NULL,
  group_filenames = F,
  ...
)
}
\arguments{
\item{ws}{path to flowjo workspace or a parsed xml-document (xml2::read_xml(ws))}

\item{lapply_fun}{function name without quotes; lapply, pbapply::pblapply or parallel::mclapply are suggested}

\item{groups}{which flowjo groups to include}

\item{group_filenames}{create a list column of filenames that have equal transformation pararmeters}

\item{...}{additional argument to the lapply function; mainly mc.cores when parallel::mclapply is chosen}
}
\value{
data frame with transformation parameters for each channel
}
\description{
Get transformation parameters from a flowjo workspace
}
\examples{
\dontrun{
trafo_df <- wsx_get_transformations(ws = ws)
}
}
