% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ab_panel_to_fcs.R
\name{ab_panel_to_fcs}
\alias{ab_panel_to_fcs}
\title{Write antibody related information into a fcs file}
\usage{
ab_panel_to_fcs(
  sampledescription,
  FileNames,
  FCS.file.folder,
  channel_conjugate_match_file = system.file("extdata", "channel_conjugate_matches.xlsx",
    package = "fcexpr"),
  AbCalcFile_col = "AbCalcFile",
  AbCalcSheet_col = "AbCalcSheet",
  AbCalcFile_folder = file.path(dirname(FCS.file.folder), "Protocols"),
  conjugate_to_desc = T,
  other_keywords = c(),
  clear_previous = F,
  ignore_duplicate_ag = F,
  machine = NULL,
  manual_df = NULL
)
}
\arguments{
\item{sampledescription}{data frame of sampledescription}

\item{FileNames}{vector of which fcs files to consider}

\item{FCS.file.folder}{path to the folder containing the fcs files specified in FileNames}

\item{channel_conjugate_match_file}{path to an xlsx file holding information about fluorochromes matched to channels (better leave as it is for now)}

\item{AbCalcFile_col}{column name in sampledescription indicating the file containing the antibody panel calculation}

\item{AbCalcSheet_col}{column name in sampledescription indicating the respective sheet name in AbCalcFile}

\item{AbCalcFile_folder}{path to the folder containing the AbCalcFile; if NULL then AbCalcFile_col must contain the full, absolute path of AbCalcFile}

\item{conjugate_to_desc}{alter/update the channel description of fcs files with stained molecule and optionally the fluorochrome}

\item{other_keywords}{column names in AbCalcSheet of which keywords to write to fcs files}

\item{clear_previous}{clear all previous entries (channel descriptions and keywords) in fcs files?}

\item{ignore_duplicate_ag}{whether or not to ignore duplicate antigen entries in antibody calculation sheet}

\item{machine}{name of the Flow Cytometer the FCS were acquired on. Only necessary when channels cannot be matched unambiguously.}

\item{manual_df}{manually provide a data frame with AbCalcFile_col, AbCalcSheet_col and FilePath columns for every FCS file
to write channel information to; sampledescription, FileNames, FCS.file.folder are not required when !is.null(manual_df)}
}
\value{
no return, but updated fcs files on disk
}
\description{
This writes channel descriptions and optional keywords to fcs files.
}
\examples{
\dontrun{
fcexpr::ab_panel_to_fcs(sampledescription = sd,
FileNames = sd$FileName,
FCS.file.folder = file.path(wd, "FCS_files"),
other_keywords = c("Isotype", "Clone", "totalDF", "Vendor", "Cat", "Lot"))
}
}
