% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/col_pal.R
\name{col_pal}
\alias{col_pal}
\title{Color palettes}
\usage{
col_pal(name = NULL, n = NULL, direction = c(1, -1))
}
\arguments{
\item{name}{name of the palette}

\item{n}{number of colors to return; may not work for every palette}

\item{direction}{reverse palette with -1}

\item{nbrew}{number of color from brewer palettes}
}
\value{
a color palette as character vector
}
\description{
Wrapper function around paletteer and two additional color palettes. See paletteer::palettes_c_names and paletteer::palettes_d_names.
}
\examples{
\dontrun{
col_pal(name = "custom", n = 10)
}
}
