% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_hd_pops.R
\name{create_pop}
\alias{create_pop}
\title{Create a random population in high dimensional space}
\usage{
create_pop(
  range_n = c(50:1000),
  dims = 5,
  range_mean = c(1:1e+05),
  range_sd_fct = c(2:10),
  range_modes = c(1:3),
  mean_prob_vec_args = list(),
  ...
)
}
\arguments{
\item{range_n}{range of random total size}

\item{dims}{number of dimensions (columns)}

\item{range_mean}{range of random population means}

\item{range_sd_fct}{population spread: mean/sample(range_sd_fct,1)}

\item{range_modes}{random number peaks (modes, modalities) per dim,
set 1 to have unimodality only, range_n is split by modes randomly
with fcexpr:::random_split_with_bounds}

\item{mean_prob_vec_args}{create a probability vector with
fcexpr::prob_vec and provide its arguments here}
}
\value{
data frame
}
\description{
Create a random population in high dimensional space
}
\examples{
# unimodal only
df <- create_pop(range_modes = 1, dims = 3)
brathering::plot3d(df, color = "group")
# trimodal
df <- create_pop(range_modes = 1:3, dims = 3)
brathering::plot3d(df, color = "group")
# tight modes
df <- create_pop(range_modes = 1:3, dims = 3, range_sd_fct = 2)
brathering::plot3d(df, color = "group")
}
