% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dimred_to_fcs.R
\name{dimred_to_fcs}
\alias{dimred_to_fcs}
\title{Calculate dimension reduction and cluster annotation with data from one or more flow frames and add these parameters to a (concatenated) fcs file}
\usage{
dimred_to_fcs(
  ff.list,
  channels = NULL,
  grouplist = NULL,
  scale.whole = c("none", "z.score", "min.max"),
  scale.samples = c("none", "z.score", "min.max"),
  run.pca = 0,
  run.umap = T,
  run.som = T,
  run.find.clusters = T,
  calc_cluster_marker = F,
  UMAP_args = list(metric = "cosine", verbose = T, scale = T),
  SOM_args = list(),
  EmbedSOM_args = list(),
  FindNeighbors_args = list(),
  FindClusters_args = list(resolution = c(0.1)),
  mc.cores = 1,
  save.to.disk = c("fcs", "rds"),
  save.path = file.path(getwd(), paste0(substr(gsub("\\\\.", "",
    make.names(as.character(Sys.time()))), 2, 15), "_FCS_dr")),
  save.name = NULL,
  write.transformed.channels.to.FCS = T,
  write.untransformed.channels.to.FCS = T,
  trafoname = "trafolistinv",
  timeChannel = c("Time", "HDR-T"),
  seed = 42,
  ...
)
}
\arguments{
\item{ff.list}{a list of flowFrames as received from fcexpr::wsp_get_ff
(compensated with Compensation Matrix as defined in FlowJo by default) or
as received from fcexpr::inds_get_ff (directly from FCS files, not
compensated by default)}

\item{channels}{channel selection, passed to ff_get_channels}

\item{grouplist}{named list of additional channels to identify samples or
group them in flowjo; see ff_add_groups}

\item{scale.whole}{if and how to scale channels after concatenation of
flowframes in ff.list}

\item{scale.samples}{if and how to scale channels of flowframes in ff.list
individually before concatenation}

\item{run.pca}{select number of pcs; if 0 no pca is computed}

\item{run.umap}{logical, whether to calculate UMAP dimension reduction
with uwot::umap}

\item{run.som}{logical, whether to calculate SOM dimension reduction
EmbedSOM::SOM followed by EmbedSOM::EmbedSOM}

\item{run.find.clusters}{logical, whether to detect clusters with
get_louvain_clusters}

\item{calc_cluster_marker}{calculate global cluster markers;
see ff_calc_markers}

\item{UMAP_args}{arguments to uwot::umap}

\item{SOM_args}{args to EmbedSOM::SOM}

\item{EmbedSOM_args}{args to EmbedSOM::EmbedSOM}

\item{FindNeighbors_args}{arguments to Seurat function fof louvain clustering}

\item{FindClusters_args}{arguments to Seurat function fof louvain clustering}

\item{mc.cores}{number of cores to use for parallel computing of
cluster markers and multiple cluster resolutions;
mc.cores is passed to parallel::mclapply or parallel::mcmapply;
limited to parallel::detectCores()-1}

\item{save.to.disk}{what to save to disk: (concatenated) and appended FCS
file and/or rds file with several elements in a list}

\item{save.path}{where to save elements specified in save.to.disk;
set to NULL to have nothing written to disk}

\item{save.name}{name to use for save rds and or fcs file}

\item{write.transformed.channels.to.FCS}{do save transformed channels
(e.g. logicle or arcsinh) to FCS file}

\item{write.untransformed.channels.to.FCS}{do save untransformed channels
(inverse) to FCS file; requires trafoname in attributes of ff}

\item{trafoname}{name of transformatio in attributes of ff}

\item{timeChannel}{name of the Time channel to exclude from all
analyses and calculation; if NULL will be attempted to be detected
automatically}

\item{seed}{set a seed for reproduction of dimension reductions}

\item{...}{arguments to ff_get_channels}
}
\value{
a list with flow frame and optionally global marker and pca results
}
\description{
It is expected that ff.list was made with wsp_get_ff with return_transformed
TRUE and return_untransformed FALSE.
After optional scaling, this runs ff_calc_pca, ff_calc_umap, ff_calc_som,
get_louvain_cluster. flow frames from ff.list are joined with ff_merge,
data from pca, umap etc added with ff_add_columns.
If write.untransformed.channels.to.FCS is TRUE, ff_transform is called.
}
\examples{
\dontrun{
ff <- wsp_get_ff(wsp = ws[18],
                 downsample = 0.05,
                 FCS.file.folder = file.path(wd, "FCS_files"),
                 groups = c("blood", "urine_spike"),
                 population = "193Ir_DNA_2, 127I_DNA_3 subset")

# select population and trafo
ffs <- purrr::map(ff, ~.x[["transformed"]][[1]])
pars <- ffs[[1]]@parameters@data
channels <-
  pars |>
  dplyr::filter(grepl("_", desc)) |>
  dplyr::filter(!grepl("Beads|DNA", desc)) |>
  dplyr::pull(name)

out <- dimred_to_fcs(
  ff.list = ffs,
  channels = channels,
  run.pca = 12,
  save.path = "/Users/vonskopnik/Downloads/",
  save.name = "testDR",
  part_match = F)
marker <- ff_calc_marker(
  out[["flowframe"]],
  cluster_cols = "res.0.1",
  channels = paste0(channels, "_trans")
)
df <- marker[["res.0.1"]][["global"]]
df <-
  df |>
  dplyr::mutate(pvalue2 = -log10(pvalue)) |>
  dplyr::mutate(pvalue2 = ifelse(is.infinite(pvalue2), 304, pvalue2))
heatmap_long_df(df = df,
                groups = "cluster",
                features = "channel",
                values = "mean_cluster_scale",
                dotsizes = "pvalue2")
}
}
