% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dr_to_fcs3.R
\name{dr_to_fcs3}
\alias{dr_to_fcs3}
\title{Calculate dimension reduction and cluster annotation with data from one or more flow frames and add these parameters to a (concatenated) fcs file}
\usage{
dr_to_fcs3(
  ff.list,
  channels = NULL,
  grouplist = NULL,
  scale.whole = c("none", "z.score", "min.max"),
  scale.samples = c("none", "z.score", "min.max"),
  run.pca = 0,
  run.umap = T,
  run.som = T,
  run.find.clusters = T,
  calc_cluster_marker = F,
  UMAP_args = list(metric = "cosine", verbose = T, scale = T),
  SOM_args = list(),
  EmbedSOM_args = list(),
  FindNeighbors_args = list(),
  FindClusters_args = list(resolution = c(0.1)),
  mc.cores = 1,
  save.to.disk = c("fcs", "rds"),
  save.path = file.path(getwd(), paste0(substr(gsub("\\\\.", "",
    make.names(as.character(Sys.time()))), 2, 15), "_FCS_dr")),
  save.name = NULL,
  write.transformed.channels.to.FCS = T,
  write.untransformed.channels.to.FCS = T,
  trafoname = "trafolistinv",
  timeChannel = c("Time", "HDR-T"),
  seed = 42,
  ...
)
}
\arguments{
\item{ff.list}{a list of flowFrames as received from fcexpr::wsp_get_ff
(compensated with Compensation Matrix as defined in FlowJo by default) or
as received from fcexpr::inds_get_ff (directly from FCS files, not
compensated by default)}

\item{channels}{channel selection, passed to ff_get_channels}

\item{grouplist}{named list of additional channels to identify samples or
group them in flowjo; see ff_add_groups}

\item{scale.whole}{if and how to scale channels after concatenation of
flowframes in ff.list}

\item{scale.samples}{if and how to scale channels of flowframes in ff.list
individually before concatenation}

\item{run.pca}{select number of pcs; if 0 no pca is computed}

\item{run.umap}{logical, whether to calculate UMAP dimension reduction
with uwot::umap}

\item{run.som}{logical, whether to calculate SOM dimension reduction
EmbedSOM::SOM followed by EmbedSOM::EmbedSOM}

\item{run.find.clusters}{logical, whether to detect clusters with
get_louvain_clusters}

\item{calc_cluster_marker}{calculate global cluster markers;
see ff_calc_markers}

\item{UMAP_args}{arguments to uwot::umap}

\item{SOM_args}{args to EmbedSOM::SOM}

\item{EmbedSOM_args}{args to EmbedSOM::EmbedSOM}

\item{FindNeighbors_args}{arugments to Seurat function fof louvain clustering}

\item{FindClusters_args}{arugments to Seurat function fof louvain clustering}

\item{mc.cores}{number of cores to use for parallel computing of
cluster markers and multiple cluster resolutions;
mc.cores is passed to parallel::mclapply or parallel::mcmapply;
limited to parallel::detectCores()-1}

\item{save.to.disk}{what to save to disk: (concatenated) and appended FCS
file and/or rds file with several elements in a list}

\item{save.path}{where to save elements specified in save.to.disk;
set to NULL to have nothing written to disk}

\item{save.name}{name to use for save rds and or fcs file}

\item{write.transformed.channels.to.FCS}{do save transformed channels
(e.g. logicle or arcsinh) to FCS file}

\item{write.untransformed.channels.to.FCS}{do save untransformed channels
(inverse) to FCS file; requires trafoname in attributes of ff}

\item{trafoname}{name of transformatio in attributes of ff}

\item{timeChannel}{name of the Time channel to exclude from all
analyses and calculation; if NULL will be attempted to be detected
automatically}

\item{seed}{set a seed for reproduction of dimension reductions}

\item{...}{arugments to ff_get_channels}
}
\value{
A list with 3 elements: (i) The matrix of fluorescence intensities and appended information (dim red, clustering). This is the table which is written into a newly generated fcs file.
(ii) A character vector of meaningful column names which may be used for the table in R (rather for convenience). (iii) Tables of marker features (each cluster vs all other events and all clusters pairwise).
}
\description{
It is expected that ff.list was made with wsp_get_ff with return_transformed
TRUE and return_untransformed FALSE.
After optional scaling, this runs ff_calc_pca, ff_calc_umap, ff_calc_som,
get_louvain_cluster. flow frames from ff.list are joined with ff_merge,
data from pca, umap etc added with ff_add_columns.
If write.untransformed.channels.to.FCS is TRUE, ff_transform is called.
}
\examples{
\dontrun{
############################
### Plot cluster markers ###
############################
dr <- fcexpr::dr_to_fcs(ff.list = ffs,
channels = channels,
louvain__resolution = 0.5,
run.lda = "louvain_0.5",
clustering.for.marker.calc = c("louvain_0.5"),
save.path = NULL)
marker <- dr[[3]][[1]][[1]]
marker$channel_desc2 <- sapply(strsplit(marker$channel_desc, "_"), "[", 1)
marker <-
 marker \%>\%
 dplyr::mutate(pvalue = ifelse(pvalue == 0, 1e-300, marker$pvalue)) \%>\%
 dplyr::group_by(channel_desc2) \%>\%
 dplyr::mutate(mean_scaled = fcexpr:::min.max.normalization(mean))

ggplot(marker, aes(x = as.factor(cluster), y = channel_desc2, fill = -log10(pvalue))) +
 geom_tile(color = "black") +
 theme_bw() +
 geom_text(aes(label = diff_sign)) +
 scale_fill_viridis_c()

ggplot(marker, aes(x = as.factor(cluster), y = channel_desc2, fill = mean_diff)) +
 geom_tile(color = "black") +
 theme_bw() +
 geom_text(aes(label = diff_sign)) +
 scale_fill_viridis_c()

ggplot(marker, aes(x = as.factor(cluster), y = channel_desc2, fill = mean_scaled)) +
 geom_tile(color = "black") +
 theme_bw() +
 scale_fill_viridis_c()

ggplot(marker, aes(x = mean_diff, y = -log10(pvalue), label = channel_desc2)) + #color = channel_desc2,
 geom_point() +
 theme_bw() +
 labs(title = "cluster markers (vs all other cells each)") +
 ggrepel::geom_text_repel(max.overlaps = 20, show.legend = F) +
 theme(panel.grid.minor.x = element_blank(), panel.grid.minor.y = element_blank(), panel.grid.major.x = element_blank(), strip.background = element_rect(fill = "hotpink2")) +
 geom_vline(xintercept = 0, col = "tomato2", linetype = "dashed") +
 geom_hline(yintercept = 100, col = "tomato2", linetype = "dashed") +
 facet_wrap(vars(cluster))

 ##############################################
 ### find clusters which may be subject #######
 ### to bi- or multimodality in any channel ###
 ##############################################
 # make one data frame
 marker_all <- purrr::map_dfr(setNames(names(out[["marker"]]), names(out[["marker"]])),
 function(x) out[["marker"]][[x]][["marker_table"]], .id = "clustering")

 # sort by diptest p value; or low p indicates bi- or multimodality
 marker_all <- dplyr::arrange(marker_all, diptest_p)
 # see ?diptest::dip.test
 # a low p-value indicates bi- or multimodality (multiple peaks)
 # a high p-value (close to 1) indicates unimodality
 # multimodality indicates heterogeneity within in the cluster
 # and may justify to separate that cluster further into sub-clusters
 # this depends on the interpretation of the scientist though


 ##############################################
 ### overlay gated populations from flowjo ####
 ### on dimension reduction plot (SOM/UMAP) ###
 ##############################################

common_cols <- Reduce(intersect, purrr::map(ff[["indices"]], colnames))
ff[["indices"]] <- purrr::map(ff[["indices"]], function(x) x[,which(colnames(x) \%in\% common_cols)])
ind_mat <- do.call(rbind, ff[["indices"]])

som <- ggplot(out[["df"]], aes(SOM_1, SOM_2, color = as.factor(cutree_30))) +
geom_point(size = 0.5) +
geom_density2d(data = . \%>\% dplyr::filter(ind_mat[,which(basename(colnames(ind_mat)) == "NK cells")]), color = "black",
contour_var = "ndensity", breaks = c(0.1, 0.2, 0.4, 0.6, 0.8)) +
guides(color = guide_legend(override.aes = list(size = 2)))

}
}
