% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fcs_get_keywords.R
\name{fcs_get_keywords}
\alias{fcs_get_keywords}
\title{Get keywords from all FCS files in a folder or subfolder}
\usage{
fcs_get_keywords(FCS.file.folder = file.path(getwd(), "FCS_files"))
}
\arguments{
\item{FCS.file.folder}{path to root folder which contains FCS files; if missing file.path(getwd(), 'FCS_files') is assumed}
}
\value{
a data frame of keywords as columns and FCS files as rows
}
\description{
The returned data frame can be used to filter for relevant keywords which may then be joined to the sampledescription.
Note that if you have defined your keywords in flowjo you have to export the FCS files in order to hard-code the keyword into the FCS file. Otherwise it only exists in the flowjo workspace. In order to pull out keywords from there another function exists.
}
\examples{
\dontrun{
key_df <- fcs_get_keywords()
key_df <- key_df[,c("FileName", "$OP")]
# sd is data frame of sampledescription
sd <- dplyr::left_join(sd, key_df, by = "FileName")
# save sd
}
}
