% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fcs_get_keywords.R
\name{fcs_get_keywords}
\alias{fcs_get_keywords}
\title{Get keywords from all FCS files in a folder or subfolder}
\usage{
fcs_get_keywords(file_path)
}
\arguments{
\item{file_path}{path to the fcs file}
}
\value{
a data frame of keywords as columns and FCS files as rows
}
\description{
The returned data frame can be used to filter for relevant keywords which may then be joined to the sampledescription.
Note that if you have defined your keywords in flowjo you have to export the FCS files in order to hard-code the keyword into the FCS file. Otherwise it only exists in the flowjo workspace. In order to pull out keywords from there another function exists.
}
\examples{
\dontrun{
# get FCS files in a folder
files <- list.files(FCS.files.folder, recursive = T, full.names = T)
# loop through files and extract keywords; only SPILL is
keys <- lapply(files, function(x) {
fcs_get_keywords(x)
})
# convert the lists of data frames to one wide data frame
keys <- tidyr::pivot_wider(do.call(rbind, keys), names_from = keyword, values_from = value)
}
}
