% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fcs_keyword_increase_max_range.R
\name{fcs_keyword_increase_max_range}
\alias{fcs_keyword_increase_max_range}
\title{Title}
\usage{
fcs_keyword_increase_max_range(
  fcs_file_path,
  channel_reg_expr = NULL,
  value = 5000
)
}
\arguments{
\item{value}{}
}
\description{
Title
}
