% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ff_add_desc.R
\name{ff_add_desc}
\alias{ff_add_desc}
\title{Add channel descriptions to flow frame}
\usage{
ff_add_desc(ff, desc)
}
\arguments{
\item{ff}{one flowframe or a list of such}

\item{desc}{named vector of channel descriptions, names must be channel names}
}
\value{
ff list
}
\description{
Add channel descriptions to flow frame
}
\examples{
\dontrun{
  ps <- fcexpr::wsx_get_popstats(w1, strip_data = F)[[1]]
  channels <- fcexpr::guess_marker_channels(ps) |>
    dplyr::filter(!ind \%in\% c(3,10,7))
  channels <- stats::setNames(channels$marker, channels$value)

  fflist <- fcexpr::wsp_get_ff(w1,
                               FCS.file.folder = file.path(wd, "FCS_files"),
                               samples = grep("full", basename(f1[[1]]), value = T),
                               population = "Time")
  ffs <- purrr::map(fflist[[1]], ~.x[[1]][[1]])
  ffs <- ff_add_desc(ffs, desc = channels)

  dr <- fcexpr::dimred_to_fcs(ffs, channels = channels)
}
}
