% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ff_calc_umap_tsne.R
\name{ff_calc_umap_tsne}
\alias{ff_calc_umap_tsne}
\title{Calculate UMAP or tSNE dimension reduction based on flow frame}
\usage{
ff_calc_umap_tsne(
  ff,
  channels = NULL,
  exprs = NULL,
  fun_args = list(n_neighbors = 20, metric = "cosine", verbose = T, scale = T),
  seed = 42,
  names = "..auto..",
  return = c("coords", "ff"),
  fun = uwot::umap,
  ...
)
}
\arguments{
\item{ff}{flow frame}

\item{channels}{which channels to use, see ff_get_channels for selection
options}

\item{exprs}{expression matrix, either provide ff or exprs}

\item{fun_args}{arguments to fun}

\item{seed}{random seed}

\item{names}{names of new dimensions to add as channels, when auto
adjusted to UMAP or tSNE with trailing dimension numbers}

\item{return}{return UMAP object or flow frame with umap channels added}

\item{fun}{uwot::umap or Rtsne::Rtsne or Seurat::RunTSNE or Seurat::RunUMAP
or brathering::fft_tsne}

\item{...}{arguments to ff_get_channels}
}
\value{
flow frame or umap object
}
\description{
Calculate UMAP or tSNE dimension reduction based on flow frame
}
\examples{
\dontrun{
mat <- hdos::random_space(n_spheres = 40, n_tori = 0, n_cuboids = 0,
                          n_samples = 2000, max_dim = 20)

umap <- ff_calc_umap_tsne(exprs = as.matrix(mat[["coord"]]),
                          fun_args = list(n_neighbors = 20, metric = "cosine", verbose = T, scale = T),
                          fun = uwot::umap)
brathering::plot2(umap)

# fast tsne: fft or fit
tsne1 <- ff_calc_umap_tsne(exprs = scale(as.matrix(mat[["coord"]])),
                           fun_args = list(),
                           fun = brathering::fft_tsne)
brathering::plot2(tsne1)

# slow tsne: barnes hut
tsne2 <- ff_calc_umap_tsne(exprs = as.matrix(mat[["coord"]])[1:4000,],
                           fun_args = list(theta = 0.8, verbose = T, pca_scale = T),
                           fun = Rtsne::Rtsne)
brathering::plot2(tsne2)

# seurat wrapper, barnes hut
tsne3 <- ff_calc_umap_tsne(exprs = as.matrix(mat[["coord"]])[1:4000,],
                           fun_args = list(theta = 0.8, verbose = T, pca_scale = T, tsne.method = "Rtsne"),
                           fun = Seurat::RunTSNE)
brathering::plot2(tsne3)

# seurat wrapper, fit
tsne4 <- ff_calc_umap_tsne(exprs = as.matrix(mat[["coord"]]),
                           fun_args = list(theta = 0.01, tsne.method = "FIt-SNE"),
                           fun = Seurat::RunTSNE)
brathering::plot2(tsne4)

## try 3D
# tsne: not supported
tsne3d <- ff_calc_umap_tsne(exprs = scale(as.matrix(mat[["coord"]])),
                            fun_args = list(dims = 3),
                            fun = brathering::fft_tsne)
# umap
umap3d <- ff_calc_umap_tsne(exprs = as.matrix(mat[["coord"]]),
                            fun_args = list(n_neighbors = 20, metric = "cosine", verbose = T, scale = T, n_components = 3),
                            fun = uwot::umap)

brathering::plot3d(umap3d)
brathering::plot3d(umap3d, backend = "rgl")
}
}
