% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ff_simulate2.R
\name{ff_simulate2}
\alias{ff_simulate2}
\title{Create a random flow frame}
\usage{
ff_simulate2(exprs = NULL, m = 1, annotate_channels = F, ...)
}
\arguments{
\item{exprs}{list of expression matrices (or data frames); if NULL then
fcexpr::make_hd_pops is used to create data}

\item{m}{number of exprs to create with make_hd_pops --> number of flow frames}

\item{...}{arguments to make_hd_pops and fcexpr::create_pop}
}
\value{
list of flow frames
}
\description{
Create a random flow frame
}
\examples{
# populations with different properties, pre-made
pops1 <- make_hd_pops(npop = 3, range_sd_fct = 20, range_modes = 2, rm_ann_cols = T)
pops2 <- make_hd_pops(npop = 3, range_sd_fct = 3:6, range_modes = 3, rm_ann_cols = T)
ff_list1 <- ff_simulate2(exprs = dplyr::bind_rows(pops1, pops2))
# have populations made within function
ff_list <- ff_simulate2(npop = 2, dims = 10, m = 3)
}
