% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gating_tree_plot.R
\name{gating_tree_plot}
\alias{gating_tree_plot}
\title{Plot gating tree as a graph to explore how populations are gated in a simple way}
\usage{
gating_tree_plot(PopulationFullPath, layout = "kk", find_short_gating_path = T)
}
\arguments{
\item{PopulationFullPath}{vector of all populations path (full) gated in a (flowjo) workspace;
optionally provide named vector (e.g. with short names) which will be plotted in the graph at correponding nodes}

\item{layout}{layout for plotting, see ?ggraph::ggraph or ?ggraph::layout_tbl_graph_igraph}

\item{find_short_gating_path}{if PopulationFullPath is a vector without names find the shortest unique
gating path from these full paths for plotting}
}
\value{
list of plot, graph and data.frame
}
\description{
Plot gating tree as a graph to explore how populations are gated in a simple way
}
\examples{
\dontrun{
# Get count data frame from flowjo wsp
ps_counts <- fcexpr::wsx_get_popstats(ws = ws)[["counts"]]
# check if gatingtree is unique
dplyr::group_by(ps_counts, PopulationFullPath) \%>\% dplyr::count()
# generate graph
graph_list <- fcexpr::gating_tree_plot(PopulationFullPath = stats::setNames(ps_counts$PopulationFullPath, ps_counts$Population))
# manual plotting graph to modify (e.g. layout)
plot <- ggraph::ggraph(graph_list[["graph"]], layout = "kk") +
ggraph::geom_edge_link() +
ggraph::geom_node_point(size = 4) +
ggraph::geom_node_label(ggplot2::aes(label = Population), repel = T)
}
}
