% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_louvain_cluster.R
\name{get_louvain_cluster}
\alias{get_louvain_cluster}
\title{Calculate louvain clusters based on expression matrix}
\usage{
get_louvain_cluster(
  exprs,
  FindNeighbors_args = list(),
  FindClusters_args = list(resolution = c(0.1)),
  mc.cores = 1
)
}
\arguments{
\item{exprs}{numeric matrix with channels as columns}

\item{FindNeighbors_args}{arguments to Seurat::FindNeighbors}

\item{FindClusters_args}{arguments to Seurat::FindClusters}

\item{mc.cores}{cores to use to calculate multiple resolutions in parallel}
}
\value{
matrix with cluster identities
}
\description{
Calculate louvain clusters based on expression matrix
}
