% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gs_get_gates.R
\name{gs_get_gates}
\alias{gs_get_gates}
\title{Gates from gatingset for plotting with ggcyto}
\usage{
gs_get_gates(
  gs,
  n_bins = 40000,
  quantile_lim_filter = c(1e-04, 0.9999),
  min_max_vals = c(0, 300),
  min_max_vals_scatter = c(0, 250000),
  x_statpos_name = 0.9,
  y_statpos_name = 0.9,
  x_statpos_pct = 0.9,
  y_statpos_pct = 0.1,
  statsize_name = 4,
  statsize_pct = 4,
  reduce_n_bins_below = 0,
  reduce_factor = 2,
  scatter_equal = T,
  fluo_equal = T
)
}
\arguments{
\item{gs}{gs a gatingset, e.g. made with fcexpr::wsp_get_gs}

\item{n_bins}{number of bins in total, will be used equally in x and y direction, bin size is adjusted to ranges in x and y direction}

\item{quantile_lim_filter}{min,max quantiles of signal intensities to set axis limits to; use quantiles to exclude extreme values}

\item{min_max_vals}{min,max required signal intensity for fluorescence channels of one event in order to consider it for axis limit calculation (to filter extreme values);
in logicle transformation}

\item{min_max_vals_scatter}{min,max required signal intensity for scatter channels of one event in order to consider it for axis limit calculation (to filter extreme values);
in inverse transformation which is equal to logicle transformation (for scatter channels)}

\item{x_statpos_name}{x-position for gate name labels}

\item{y_statpos_name}{y-position for gate name labels}

\item{x_statpos_pct}{x-position for gate percent labels}

\item{y_statpos_pct}{y-position for gate percent labels}

\item{statsize_name}{size of name label}

\item{statsize_pct}{size of percent label}

\item{reduce_n_bins_below}{reduce number of bins when number of events is below this number}

\item{reduce_factor}{by which factor to reduce bin number?}

\item{scatter_equal}{equal limits for all scatter channels?}

\item{fluo_equal}{equal limits for each fluorescence channel in all panels?}
}
\value{
a data frame to loop over and produce plots with ggcyto
}
\description{
Gates from gatingset for plotting with ggcyto
}
\examples{
\dontrun{
gates <- gs_get_gates(gs)
}
}
