% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gslist_try_merge.R
\name{gslist_try_merge}
\alias{gslist_try_merge}
\title{Try to merge gatingsets by removing non-common populations/gates}
\usage{
gslist_try_merge(gs_list)
}
\arguments{
\item{gs_list}{list of gatingsets}
}
\value{
gatingset
}
\description{
Uses flowWorkspace::merge_list_to_gs
}
\examples{
\dontrun{
# read gatingsets
gslist1 <- wsp_get_gs(wsp1)
gslist2 <- wsp_get_gs(wsp2)
gs <- gslist_try_merge(list(gslist1[[1]][[1]], gslist2[[1]][[1]]))
}
}
