% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatmap_long_df.R
\name{heatmap_long_df}
\alias{heatmap_long_df}
\title{Plot heatmap from data frame in long format}
\usage{
heatmap_long_df(
  df,
  groups,
  features,
  values,
  featuregroup = NULL,
  featuregroup_style = c("facet", "color"),
  featuregroup_col_name = "",
  featuregroup_col_pal = "custom_light",
  dotsizes = NULL,
  dotsize_range = c(2, 7),
  fill = "..auto..",
  color = "..auto..",
  scale = c("none", "zscore", "1"),
  features_topn = NULL,
  topn_cols = values,
  topn_ties = F,
  featurelabels = "..auto..",
  featurelabels_repel = F,
  featuresitalic = F,
  color_linewidth = 0.2,
  legendbreaks = "..auto..",
  legendlabels = "..auto..",
  colorsteps = "..auto..",
  colorsteps_nice = T,
  axes_flip = F,
  group_seplines = F,
  seplines_args = list(),
  theme = ggplot2::theme_classic(),
  legend_fill_args = list(barwidth = 1, barheight = 8, order = 1),
  legend_size_args = list(order = 2, ncol = NULL, nrow = NULL, override.aes = list(color
    = "..auto..", fill = "..auto..")),
  theme_args = list(panel.grid = ggplot2::element_blank()),
  repel_args = list(featurelabels_width = 0.2, featurelabels_nudhe_x = -1),
  heatmap_ordering_args = list(feature_order = "custom", group_order = "hclust"),
  values_zscored = NULL,
  ...
)
}
\arguments{
\item{df}{data frame in long format}

\item{groups}{column with groups}

\item{features}{column with features}

\item{values}{column with fill values for color scale}

\item{featuregroup}{column name in df that groups features}

\item{featuregroup_style}{how to show feature groups, by colored axis text and/or separate facets}

\item{featuregroup_col_name}{color legend name}

\item{featuregroup_col_pal}{color palette name passed to colrr::col_pal}

\item{dotsizes}{column for dot sizes, if provided will cause plotting of
dots instead of tiles}

\item{dotsize_range}{range for dot size}

\item{fill}{color palette vector for fill of tiles or dots. when auto,
RColorBrewer::RdBu is used}

\item{color}{color of stroke (border) around tiles or dots; "NA" means no
stroke is plotted; NA has
to be put in quotation mark ("NA"), such that geom_point accepts it.
other choices may be black, white or any other color code; when "auto",
grey70 is used by default when is.null(dotsizes) and a the number of
features is below 100.}

\item{scale}{how to scale values: not, zscore or from -1 to 1}

\item{features_topn}{only plot the top n features (ordered by value) per
group}

\item{topn_cols}{columns in df to use for ordering when selecting
features_topn with dplyr::slice_max. hint: p-values have to multiplied
by -1 to make this work properly}

\item{topn_ties}{break ties for features_topn? if TRUE, more then
features_topn may be plotted}

\item{featurelabels}{subset of feature labels to plot; NULL to plot all,
"" to plot none; can be a named vector with names being aliases to use for
plotting e.g. c("CD20" = "MS4A1", "CD3", "KLRG1") to only alter MS4A1;
auto to omit labels by default when more than 100 features are there}

\item{featurelabels_repel}{do repel feature axis labels?}

\item{featuresitalic}{shorthand to plot feature labels in italic}

\item{color_linewidth}{linewidth of borders (stroke) around tiles or dots}

\item{legendbreaks}{a single number, a vector of explicit breaks, or "auto"
for ggplot default or "minmidmax" for three breaks at minimum, middle and
maximum of value range}

\item{legendlabels}{labels for breaks, e.g. c("min", "mid", "max")}

\item{colorsteps}{NULL to have normal colorbar, auto for default colorsteps,
a single number or a vector of explicit steps; may not work with any number
when colorsteps_nice is TRUE}

\item{colorsteps_nice}{heuristic for pretty steps}

\item{axes_flip}{do flip them?}

\item{group_seplines}{plot lines that separate features belonging to
different groups}

\item{seplines_args}{arguments to geom_hline for seplines}

\item{theme}{ggplot2 theme}

\item{legend_fill_args}{arguments to ggplot2::guide_colorsteps or
ggplot2::guide_colorbar, depend upon the color scale}

\item{legend_size_args}{arguments to ggplot2::guide_legend to modify the
size legend; e.g. use override.aes = list(size = c(1,3,5)) to adjust dot size
in legend in contrast to dotsize_range, one number for each dot in legend
needed}

\item{theme_args}{arguments to ggplot2::theme}

\item{repel_args}{fine tuning for featurelabel repelling}

\item{heatmap_ordering_args}{arguments to heatmap_ordering like feature_order or group_order}

\item{values_zscored}{say if values are z-scored; if NULL inferred from data
with brathering::is_z_scored}

\item{...}{arguments to heatmap_ordering like feature_order or group_order}
}
\value{
ggplot2 object
}
\description{
Plot heatmap from data frame in long format
}
\examples{
df <- readRDS(system.file("extdata", "heatmap_df.rds", package = "fcexpr"))
# everything default
fcexpr::heatmap_long_df(df = df,
                groups = "cluster",
                features = "channel",
                values = "mean_cluster_scale")
# -log10(pvalues) as dot size
fcexpr::heatmap_long_df(df = df,
                groups = "cluster",
                features = "channel",
                values = "mean_cluster_scale",
                dotsizes = "pvalue2")
# continuuos color scale, no feature axis text, flipped axes
# only 4 features per group, lines to separate features
fcexpr::heatmap_long_df(df = df,
                groups = "cluster",
                features = "channel",
                values = "mean_cluster_scale",
                dotsizes = "pvalue2",
                featurelabels_omit = T,
                axes_flip = T,
                features_topn = 4,
                group_seplines = T,
                colorsteps = NULL)
# scale in function, alter legend text
fcexpr::heatmap_long_df(df = df,
                groups = "cluster",
                features = "channel",
                values = "mean_cluster",
                scale = "zscore",
                colorsteps = NULL,
                legendlabels = c("min", "mid", "max"),
                legendbreaks = "minmidmax")
}
