% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatmap_long_df.R
\name{heatmap_long_df}
\alias{heatmap_long_df}
\title{Plot heatmap from data frame in long format}
\usage{
heatmap_long_df(
  df,
  groups,
  features,
  values,
  dotsizes = NULL,
  fill = "auto",
  color = "auto",
  scale = c("none", "zscore", "1"),
  featurelabels_omit = F,
  features_topn = NULL,
  color_linewidth = 0.2,
  legendbreaks = "auto",
  legendlabels = "auto",
  colorsteps = "auto",
  colorsteps_nice = T,
  axes_flip = F,
  group_seplines = F,
  seplines_args = list(),
  legend_fill_args = list(barwidth = 1, barheight = 8, order = 1),
  legend_size_args = list(order = 2, ncol = NULL, nrow = NULL, override.aes = list(color
    = "black")),
  ...
)
}
\arguments{
\item{df}{data frame in long format}

\item{groups}{column with groups}

\item{features}{column with features}

\item{values}{column with fill values for color scale}

\item{dotsizes}{column for dot sizes, if provided will cause plotting of
dots instead of tiles}

\item{fill}{color palette vector for fill of tiles or dots. when auto,
RColorBrewer::RdBu is used}

\item{color}{color of stroke (border) around tiles or dots; "NA" means no
stroke is plotted; NA has
to be put in quotation mark ("NA"), such that geom_point accepts it.
other choices may be black, white or any other color code; when "auto",
grey70 is used by default when !is.null(dotsizes) and a the number of
features is below 100.}

\item{scale}{how to scale values: not, zscore or from -1 to 1}

\item{featurelabels_omit}{do not plot axis text of features, e.g. when
their number is large}

\item{features_topn}{only plot the top n features (ordered by value) per
group}

\item{color_linewidth}{linewidth of borders (stroke) around tiles or dots}

\item{legendbreaks}{a single number, a vector of explicit breaks, or "auto"
for ggplot default or "minmidmax" for three breaks at minimum, middle and
maximum of value range}

\item{legendlabels}{labels for breaks, e.g. c("min", "mid", "max")}

\item{colorsteps}{NULL to have normal colorbar, auto for default colorsteps,
a single number or a vector of explicit steps; may not work with any number
when colorsteps_nice is TRUE}

\item{colorsteps_nice}{heuristic for pretty steps}

\item{axes_flip}{do flip them?}

\item{group_seplines}{plot lines that separate features belonging to
different groups}

\item{seplines_args}{arguments to geom_hline for seplines}

\item{legend_fill_args}{arguments to ggplot2::guide_colorsteps or
ggplot2::guide_colorbar, depend upon the color scale}

\item{legend_size_args}{arguments to ggplot2::guide_legend to modify the
size legend}

\item{...}{arguments to heatmap_ordering}
}
\value{
ggplot2 object
}
\description{
Plot heatmap from data frame in long format
}
\examples{
df <- readRDS(system.file("extdata", "heatmap_df.rds", package = "fcexpr"))
# everything default
fcexpr::heatmap_long_df(df = df,
                groups = "cluster",
                features = "channel",
                values = "mean_cluster_scale")
# -log10(pvalues) as dot size
fcexpr::heatmap_long_df(df = df,
                groups = "cluster",
                features = "channel",
                values = "mean_cluster_scale",
                dotsizes = "pvalue2")
# continuuos color scale, no feature axis text, flipped axes
# only 4 features per group, lines to separate features
fcexpr::heatmap_long_df(df = df,
                groups = "cluster",
                features = "channel",
                values = "mean_cluster_scale",
                dotsizes = "pvalue2",
                featurelabels_omit = T,
                axes_flip = T,
                features_topn = 4,
                group_seplines = T,
                colorsteps = NULL)
# scale in function, alter legend text
fcexpr::heatmap_long_df(df = df,
                groups = "cluster",
                features = "channel",
                values = "mean_cluster",
                scale = "zscore",
                colorsteps = NULL,
                legendlabels = c("min", "mid", "max"),
                legendbreaks = "minmidmax")
}
