% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inds_get_ff.R
\name{inds_get_ff}
\alias{inds_get_ff}
\title{Get (subsetted) flowFrames from FCS files}
\usage{
inds_get_ff(
  ind_mat,
  population,
  alias_attr_name = "short_names",
  path_attr_name = "FilePath",
  downsample = 1,
  return_untransformed = T,
  return_logicle_transformed = T,
  lapply_fun = lapply,
  ...
)
}
\arguments{
\item{ind_mat}{a list of indices matrices, preferentially generated with fcexpr::wsp_get_indices}

\item{population}{which population (=node, =gate) to subset flowFrames on; must be a column name of ind_mat or an alias stored in alias_attr_name}

\item{alias_attr_name}{the name of attribute containing aliases (shortest unique names) of node-names (gating paths)}

\item{path_attr_name}{the name of attribute containing the path (url) to the fcs file on which to apply the subsetting}

\item{downsample}{numeric, if < 0 then a fraction of events is sampled, if > 0 an absolute number of events is sampled; or set to "min"
which will lead to downsampling each flowframe to the number of events in the flowframe with lowest number of events; can be a single value to treat all
FCS files equally or can be a vector of same length as FCS files}

\item{return_untransformed}{logical; do return untransformed (inverse) data}

\item{return_logicle_transformed}{logical; do return logicle-transformed data}

\item{lapply_fun}{lapply function name, unquoted; lapply, pbapply::pblapply or parallel::mclapply are suggested}

\item{...}{additional argument to the lapply function; mainly mc.cores when parallel::mclapply is chosen as lapply_fun}
}
\value{
list of flow frames, one for each ind_mat
}
\description{
From a matrix of indices generated with fcexpr::wsp_get_indices flowframes are generated.
Only those events within the selected population are included in flowframes. By default
no compensation will be applied on fluorescence intensites. This can be done afterwards
though with in appropriate compensation matrix.
}
\examples{
\dontrun{
ind_mat <- fcexpr::wsp_get_indices("mypath/my.wsp")
ff <- inds_get_ff(ind_mat = ind_mat, population = "CD8+")
}
}
