% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_exp.R
\name{new_exp}
\alias{new_exp}
\title{Title}
\usage{
new_exp(path = NULL, name = NULL, date_prefix = T)
}
\arguments{
\item{path}{character}

\item{name}{character}

\item{date_prefix}{logical}
}
\value{
No return value. Instead a folder is written to disk.
}
\description{
Title
}
\examples{
\dontrun{
new_exp(path = '/Users/CMS/Documents/experiments', name = 'CD3_titration', date_prefix = T)
}
}
