% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_gates_default.R
\name{plot_gates}
\alias{plot_gates}
\title{Plot a default layout of gated populations of a set of FCS files}
\usage{
plot_gates(
  gs,
  gates_df,
  facetting = NULL,
  plot_gates = T,
  plot_gate_names = T,
  plot_gate_pct = T,
  inverse_trans = F,
  geom = c("hex", "pointdensity", "scattermore"),
  gate_color = "black",
  gate_stats_color = "black",
  pct_digits = 1,
  plot_contours = F,
  plot_title = T,
  title = c("full_path", "short_path", "final_node"),
  contour_args = list(fill = "white", geom = "polygon", color = "black", contour_var =
    "ndensity", breaks = seq(0.05, 0.95, 0.1), alpha = 0.8, linewidth = 0.2),
  col_pal = colrr::col_pal("Spectral", direction = -1),
  col_pal_trans = "pseudo_log",
  theme = ggplot2::theme_bw(),
  theme_args = list(panel.grid = ggplot2::element_blank(), strip.background =
    ggplot2::element_rect(fill = "grey95", color = "white"), axis.text =
    ggplot2::element_blank(), axis.ticks = ggplot2::element_blank(), axis.title.x =
    ggplot2::element_text(margin = ggplot2::margin(t = 2, unit = "pt")), axis.title.y =
    ggplot2::element_text(margin = ggplot2::margin(r = 2, unit = "pt")), legend.position
    = "none", strip.text.x = ggplot2::element_text(margin = ggplot2::margin(2, 0, 2, 0,
    unit = "pt")), strip.text.y = ggplot2::element_text(margin = ggplot2::margin(0, 
    
    2, 0, 2, unit = "pt")), plot.margin = grid::unit(c(1, 1, 1, 1), "pt"), panel.spacing
    = grid::unit(2, "pt")),
  max_nrow_to_plot = switch(geom, hex = 50000, pointdensity = 2000, scattermore = 2e+06),
  geom_args = list()
)
}
\arguments{
\item{gs}{a gatingset, e.g. made with fcexpr::wsp_get_gs}

\item{gates_df}{a data frame with information of how to plot gates,
made with fcexpr::gs_get_gates}

\item{facetting}{which facetting to apply, facet_wrap or facet_grid
with respective arguments, check flowCore::pData(gs) for valid columns;
facet_null will completely remove facets; by default facetting is done across
each fcs file}

\item{plot_gates}{logical whether to plot gates}

\item{plot_gate_names}{logical whether to plot gate names}

\item{plot_gate_pct}{logical whether to plot gate percentages
(fraction of parent)}

\item{inverse_trans}{logical whether to inverse transform axes numbers
if TRUE this will make axes look like in flowjo}

\item{geom}{how to plot data; recommendation is hex; hex = geom_hex taking
the binwidths column of gates_df into account,
pointdensity = ggpointdensity::geom_pointdensity
scattermore = scattermore::geom_scattermore (only black-white currently)}

\item{gate_color}{line color of gates}

\item{gate_stats_color}{font color of gate statistics}

\item{pct_digits}{how many digits after comma to print;
passed to 'digits' of ggcyto::geom_stats}

\item{plot_contours}{logical whether to plot contour lines on top with
ggplot2::geom_density_2d}

\item{plot_title}{do plot titles?}

\item{title}{title format}

\item{contour_args}{arguments to stat_density_2d}

\item{col_pal}{color palette to use for color gradient generation}

\item{col_pal_trans}{argument passed as 'trans' in scale_fill_gradientn}

\item{theme}{ggplot theme}

\item{theme_args}{arguments to ggplot2::theme}

\item{max_nrow_to_plot}{to ggcyto::ggcyto; defaults are chosen with respect
to plotting time}

\item{geom_args}{arguments to geom}
}
\value{
a list of ggplot2 objects, one for every gating level;
each list index contains respective plots for every fcs file
}
\description{
Getting an overview of the results of a flow cytometry experiment can be a
generic task if gating and analysis strategy is fixed.
Here, ggcyto is used to create an arrangement of pseudocolor plots. A number
of default settings has been selected which may guarantee
good looking plots for most cases. Upon problems, the function may be
improved to handle edge cases. Currently, customization is limited to a few
rather technical aspects of the plots. This may be subject to expansion but
generally the range of customization is too large to fit it into a generic
function with a finite number of arguments. Hence, for very specific and
detailed requirements manual plotting and fiddling is unavoidable. For that,
the code of this function may serve as a template.
}
\examples{
\dontrun{
## read gatingset
gs <- fcexpr::wsp_get_gs(wsp = ws, groups = "Group1")

## write meta data to pData of gs; sd is sampledescription
p.df <-
flowCore::pData(gs) \%>\%
tibble::rownames_to_column("FileName") \%>\%
dplyr::left_join(sd) \%>\%
tibble::column_to_rownames("FileName")
p.df$FileName <- rownames(p.df)
flowCore::pData(gs) <- p.df

## get the gates_df, optionally select relevant gates, and modify
gates <- fcexpr::gs_get_gates(gs, n_bins = 100^2)
gates <- gates[which(gates$Population \%in\% c("CD8+", "CD8-")),]
gates$facet_strip <- T

## selected gates; to order facetted plot the inline factor level
## ordering is required as flowCore::pData(gs) cannot contain factors
## axis.text = element_blank() is part of ... and will omit axis numbers (passed to ggplot2::theme)
plotlist <-
fcexpr::plot_gates(gs = gs,
gates_df = gates,
facetting = facet_grid(cols = vars(factor(dilution_factor, levels = c(unique(p.df$dilution_factor)))), rows = vars(CD8_biotin_batch)),
axis.text = element_blank())

## paste plots together with patchwork and save
## patchwork is superior to cowplot as is will completely ignore ommitted facet_strips
ggsave(patchwork::wrap_plots(plotlist, ncol = 1), filename = paste0("facsplots.png"), device = "png", path = im_path, dpi = "retina", width = 18, height = 7)
}
}
