% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_gates_default.R
\name{plot_gates}
\alias{plot_gates}
\title{Plot a default layout of gated populations of a set of FCS files (e.g. from one experiment)}
\usage{
plot_gates(
  gs,
  gates_df,
  facetting = NULL,
  plot_gates = T,
  plot_gate_names = T,
  plot_gate_pct = T,
  ...
)
}
\arguments{
\item{gs}{a gatingset, e.g. made with fcexpr::wsp_get_gs}

\item{gates_df}{a data frame with informaion of how to plot gates, made with fcexpr::gs_get_gates}

\item{facetting}{which facetting (ggplot2) to apply, facet_wrap or facet_grid with respective arguments, check flowCore::pData(gs) for valid columns}

\item{plot_gates}{logical whether to plot gates}

\item{plot_gate_names}{logical whether to plot gate names}

\item{plot_gate_pct}{logical whether to plot gate percentages (fraction of parent)}

\item{...}{arguments passed to ggplot2::theme; set a global default theme with ggplot2:theme_set() and ggplot2::theme_update()}
}
\value{

}
\description{
Plot a default layout of gated populations of a set of FCS files (e.g. from one experiment)
}
