% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampledescription_to_fcs_keywords.R
\name{sampledescription_to_fcs_keywords}
\alias{sampledescription_to_fcs_keywords}
\title{Write columns from sampledescription as keyword into FCS files}
\usage{
sampledescription_to_fcs_keywords(sampledescription, columns, FCS.file.folder)
}
\arguments{
\item{sampledescription}{data frame}

\item{columns}{vector of column names from sampledescription to write as keywords}

\item{FCS.file.folder}{path to folder of FCS files}
}
\value{
no return but keywords written to FCS files
}
\description{
Given a sampledescription with annotation of fcs files such meta data may be
written as keyword into respective fcs files.
}
\examples{
\dontrun{
wd <- dirname(dirname(rstudioapi::getActiveDocumentContext()$path))
sd <- openxlsx::read.xlsx(file.path(wd, sampledescription.xlsx))
# if only a subset of files should be considered, select respective rows
sampledescription_to_fcs_keywords(sampledescription = sd, columns = c("Patient", "ExpPart"), FCS.file.folder = file.path(wd, "FCS_files"))
}
}
