% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sync_sampledescription.R
\name{sync_sampledescription}
\alias{sync_sampledescription}
\title{Title}
\usage{
sync_sampledescription(
  wd,
  FCS.file.folder,
  xlsx.file.name = "sampledescription.xlsx",
  sample.sheet.name = "samples",
  exclude.folders = c("compensation", "other_fcs_files", "experiment.file",
    "deleted_fcs_files")
)
}
\arguments{
\item{wd}{character}

\item{FCS.file.folder}{character}

\item{xlsx.file.name}{character}

\item{sample.sheet.name}{character}

\item{exclude.folders}{character}
}
\value{
No return value. Instead a sampledescription and FCS files are synced.
}
\description{
Title
}
