% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sync_sampledescription.R
\name{sync_sampledescription}
\alias{sync_sampledescription}
\title{Have an excel file (sampledescription) created or modified to store meta data of FCS files in}
\usage{
sync_sampledescription(
  FCS.file.folder,
  xlsx.file.name = "sampledescription.xlsx",
  exclude.folders = c("compensation", "other_fcs_files", "experiment.file",
    "deleted_fcs_files"),
  init.columns = c("AbCalcFile", "AbCalcSheet", "ExpProtocolFile", "ExpPart"),
  write.log = T
)
}
\arguments{
\item{FCS.file.folder}{path to the root folder which contains FCS files}

\item{xlsx.file.name}{name of the sampledescription file}

\item{exclude.folders}{character vector of folders to exclude when checking for FCS files}

\item{init.columns}{additional columns to add to the initial file}

\item{write.log}{write a hidden log file to track back changes (currently only on Mac)}
}
\value{
No return value. Instead sampledescription.xlsx and FCS files are synced.
}
\description{
Four cases can be handled: (i) if no xlsx-file named xlsx.file.name exists
in wd a xlsx-file is initiated based on the FCS files in FCS.file.folder. (ii)
When new FCS files are added to FCS.file.folder and the function is called
they are added in order of acquisition to the xlsx-file. (iii) When file names
in the FileName column of the xlsx-file are changed and the function is called
the FCS files in FCS.file.folder are renamed accordingly. (iv) If FCS files
are to be excluded or removed the entry in the FileName column has to be left
blank and the function has to be called.
}
\details{
Attention: Close the xlsx-file before calling the function so the file can be edited.
Never edit the 'identity' columns in the xlsx-file manually.
}
\examples{
\dontrun{
# When the script is saved to R_scripts in the experiment folder, this will
write the path of the experiment folder into wd:
wd <- dirname(dirname(rstudioapi::getActiveDocumentContext()$path))
# If the sampledescription is to be initiated call
sync_sampledescription(FCS.file.folder = file.path(wd, "FCS_files"))
}
}
