% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sync_sampledescription.R
\name{sync_sampledescription}
\alias{sync_sampledescription}
\title{Synchronize an excel file (sampledescription.xlsx) with FCS files}
\usage{
sync_sampledescription(
  FCS.file.folder,
  file.name = "sampledescription.xlsx",
  exclude.folders = c("compensation", "other_fcs_files", "experiment.file",
    "deleted_fcs_files"),
  init.columns = c("AbCalcFile", "AbCalcSheet", "ExpProtocolFile", "ExpPart"),
  write.log = T
)
}
\arguments{
\item{FCS.file.folder}{path to the folder which contains FCS files}

\item{file.name}{name of the sampledescription file, one of the following file types: .xlsx, .ods, .txt, .tsv, .csv}

\item{exclude.folders}{character vector of folders to exclude when checking for FCS files}

\item{init.columns}{additional columns to add to the initial file}

\item{write.log}{write a hidden (not hidden on windows) log file every time changes take place}
}
\value{
No return value. Instead sampledescription.xlsx and FCS files are synced.
}
\description{
Five cases can be handled: (i) if no xlsx-file named file.name exists
an xlsx-file is initiated based on the FCS files in FCS.file.folder. (ii)
When new FCS files are added to FCS.file.folder and the function is called
they are added in order of acquisition to the xlsx-file. (iii) When file names
in the FileName column of the xlsx-file are altered and the function is called
the FCS files in FCS.file.folder are renamed accordingly. (iv) If FCS files
are to be excluded or removed, the entry in the FileName column has to be left
blank and the function has to be called. (v) When the order of rows in the xlsx-file
is changed prefixes will be re-numbered.
}
\details{
The xlsx-file will always be written to the root of the FCS.file.folder.
Preferentially, do not have the xlsx-files open in excel when calling the the function.
Never edit the 'identity' column in the xlsx-file manually.
Never mix up rows of FileName and identity.
The identity column contain a concatenated string of the $FIL keyword, the number of events
and the acquisition date time of the FCS file.
}
\examples{
\dontrun{
sync_sampledescription(FCS.file.folder = file.path(wd, 'FCS_files'))
}
}
