% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ws_get_popstats.R
\name{ws_get_popstats}
\alias{ws_get_popstats}
\title{Title}
\usage{
ws_get_popstats(ws, gr, FCS.file.folder, groupwise = T)
}
\arguments{
\item{ws}{character}

\item{gr}{character}

\item{FCS.file.folder}{character}

\item{groupwise}{logical}
}
\value{
data.frame
}
\description{
Title
}
\examples{
\dontrun{
ws_get_popstats(ws = 'fj.wsp', gr = 'full.stain', FCS.file.folder = 'FCS.files', groupwise = T)
}
}
