% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wsp_get_ff.R
\name{wsp_get_ff}
\alias{wsp_get_ff}
\title{Get (subsetted) flowFrames from FCS files in one or many flowjo workspaces}
\usage{
wsp_get_ff(
  wsp,
  FCS.file.folder = NULL,
  groups = NULL,
  population,
  invert_groups = F,
  samples = NULL,
  invert_samples = F,
  inverse_transform = c(T, F),
  downsample = 1,
  remove_redundant_channels = F,
  lapply_fun = lapply,
  ...
)
}
\arguments{
\item{wsp}{vector of paths to flowjo workspaces}

\item{FCS.file.folder}{path to folder(s) of FCS files; may be one path for all wsp or a vector of paths, one for each wsp;
if not provided fcs file paths are derived individually from the wsp (xml)}

\item{groups}{vector or list of groups in flowjo to consider; if a list, each index corresponds to the index in wsp;
if NULL samples from all groups are read}

\item{population}{which population (=node, =gate) to subset flowFrames one; use wsx_get_poppaths to get paths}

\item{invert_groups}{logical whether to invert group selection}

\item{samples}{vector or list of samples to select (names of FCS files), each index corresponds to the index in wsp;
if NULL all samples (from selected groups) are read}

\item{invert_samples}{logical whether to invert sample selection}

\item{inverse_transform}{return inverse- (T) or logicle- (F) transform or both (c(T,F))}

\item{downsample}{numeric, if < 0 then a fraction of each flowFrame is sampled, if > 0 an absolute number of each flowFrame is subsetted}

\item{remove_redundant_channels}{remove channels that are not part of the gating tree, mainly to reduce memory load}

\item{lapply_fun}{lapply function name, unquoted; lapply, pbapply::pblapply or parallel::mclapply are suggested}

\item{...}{additional argument to the lapply function; mainly mc.cores when parallel::mclapply is chosen}
}
\value{
a list of flowframes
}
\description{
Get (subsetted) flowFrames from FCS files in one or many flowjo workspaces
}
