% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wsp_get_gs.R
\name{wsp_get_gs}
\alias{wsp_get_gs}
\title{Get (subsetted) gatingsets from flowjo workspaces}
\usage{
wsp_get_gs(
  wsp,
  FCS.file.folder = NULL,
  groups = NULL,
  invert_groups = F,
  samples = NULL,
  invert_samples = F,
  remove_redundant_channels = F,
  pData = NULL,
  pData_join_col = "identity",
  get_gates = T,
  get_gates_args = list(n_bins = 200^2),
  force_gs_merge = F
)
}
\arguments{
\item{wsp}{vector of paths to flowjo workspaces}

\item{FCS.file.folder}{path to folder(s) of FCS files; may be one path for all wsp or a vector of paths, one for each wsp;
if not provided (NULL) fcs file paths are derived individually from wsps (xml)}

\item{groups}{vector or list of groups in flowjo to consider; if a list, each index corresponds to the index in wsp;
if NULL samples from all groups are read}

\item{invert_groups}{logical whether to invert group selection}

\item{samples}{vector or list of samples to select (names of FCS files), each index corresponds to the index in wsp;
if NULL all samples (from selected groups) are read}

\item{invert_samples}{logical whether to invert sample selection}

\item{remove_redundant_channels}{remove channels that are not part of the gating tree, mainly to reduce memory load}

\item{pData}{data frame to join via rownames of pData(gs), this will make meta
data columns available for plotting with fcexpr::plot_gates}

\item{pData_join_col}{column name in pData for joining; joinable columns
are found with fcexpr:::find_unique_level_columns(pData, pData_join_col)}

\item{get_gates}{run fcexpr::gs_get_gates and include results in the return list}

\item{get_gates_args}{argument list for get_gates}

\item{force_gs_merge}{try to eliminate non-congruent populations from gating
hierarchies to allow merging of all samples into one gs}
}
\value{
list of gatingsets, FCS files used, gating hierarchies
}
\description{
Wsp files are checked for corresponding gating hierarchies and FCS files are
read into separate GatingSets (gs), respectively. That means FCS files from different
wsp are combined if their gatings are equal. But simple import from one wsp is
also possible. If file paths to FCS files in flowjo are not valid on the current computer
(e.g. because flowjo analysis was done on another computer), you have to provide the
local top folder which contains all of the necessary FCS files: FCS.files.folder.
Therein, the files are searched for recursively. Samples from wsp(s) with missing FCS files are skipped.
}
\examples{
\dontrun{
# check fcs files in wsp first
samples <- fcexpr::wsx_get_fcs_paths(ws = "mypath/my.wsp", split = F, filter_AllSamples = T)
# read-in Gating set with selected samples
gs_list <- fcexpr::wsp_get_gs(wsp = "mypath/my.wsp", FCS.files.folder = "myLocalTopFolder/subfolder", groups = "Compensation", invert_groups = T, samples = samples$FileName[1:5])
}
}
