% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wsp_get_gs.R
\name{wsp_get_gs}
\alias{wsp_get_gs}
\title{Get (subsetted) gatingsets from flowjo workspaces}
\usage{
wsp_get_gs(
  wsp,
  FCS.file.folder = NULL,
  groups = NULL,
  invert_groups = F,
  samples = NULL,
  invert_samples = F,
  remove_redundant_channels = F
)
}
\arguments{
\item{wsp}{vector of paths to flowjo workspaces}

\item{FCS.file.folder}{path to folder(s) of FCS files; may be one path for all wsp or a vector of paths, one for each wsp;
if not provided (NULL) fcs file paths are derived individually from wsps (xml)}

\item{groups}{vector or list of groups in flowjo to consider; if a list, each index corresponds to the index in wsp;
if NULL samples from all groups are read}

\item{invert_groups}{logical whether to invert group selection}

\item{samples}{vector or list of samples to select (names of FCS files), each index corresponds to the index in wsp;
if NULL all samples (from selected groups) are read}

\item{invert_samples}{logical whether to invert sample selection}

\item{remove_redundant_channels}{remove channels that are not part of the gating tree, mainly to reduce memory load}
}
\value{
list of gatingsets
}
\description{
Get (subsetted) gatingsets from flowjo workspaces
}
\examples{
\dontrun{
gs_list <- fcexpr::wsp_get_gs(wsp = "mypath/my.wsp")
}
}
