% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wsx_get_fcs_paths.R
\name{wsx_get_fcs_paths}
\alias{wsx_get_fcs_paths}
\title{Title}
\usage{
wsx_get_fcs_paths(ws)
}
\arguments{
\item{ws}{path to flowjo workspace or a parsed xml-document (xml2::read_xml(ws))}
}
\value{
a list of paths to fcs files grouped by groups in the ws
}
\description{
Title
}
\examples{
\dontrun{
wsx_get_fcs_paths(ws)
}
}
