% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wsx_get_fcs_paths.R
\name{wsx_get_fcs_paths}
\alias{wsx_get_fcs_paths}
\title{Get file paths to the fcs files in a workspace}
\usage{
wsx_get_fcs_paths(ws)
}
\arguments{
\item{ws}{path to flowjo workspace or a parsed xml-document (xml2::read_xml(ws))}
}
\value{
a list of paths to fcs files grouped by groups in the ws
}
\description{
If one has a large experiment with many fcs files and many workspaces with various
gatings finding out which fcs files are used in which workspace may be helpful.
}
\examples{
\dontrun{
wsx_get_fcs_paths(ws)
}
}
