% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wsx_get_groups.R
\name{wsx_get_groups}
\alias{wsx_get_groups}
\title{Retrieve groups within a flowjo workspace and associated samples (sampleID)}
\usage{
wsx_get_groups(ws, filter_AllSamples = T)
}
\arguments{
\item{ws}{path to flowjo workspace or a parsed xml-document (xml2::read_xml(ws))}

\item{filter_AllSamples}{filter the All Samples Group if the fcs file is also part of another group}
}
\value{
a data frame
}
\description{
Retrieve groups within a flowjo workspace and associated samples (sampleID)
}
