% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wsx_get_popstats.R
\name{wsx_get_popstats}
\alias{wsx_get_popstats}
\title{Read data from a flowjo wsp file}
\usage{
wsx_get_popstats(
  ws,
  groups = NULL,
  invert_groups = F,
  return_stats = F,
  lapply_fun = lapply,
  strip_data = T,
  skip_legacy = F,
  ...
)
}
\arguments{
\item{ws}{path to flowjo workspace or a parsed xml-document (xml2::read_xml(ws))}

\item{groups}{vector of flowjo group names to consider}

\item{invert_groups}{logical whether to exclude the selected groups}

\item{return_stats}{logical whether to return statistics}

\item{lapply_fun}{function name without quotes; lapply, pbapply::pblapply
or parallel::mclapply are suggested; only for legacy method}

\item{strip_data}{remove extensive internal data of wsp from count data frame}

\item{skip_legacy}{do not run the legacy method?}

\item{...}{additional argument to the lapply function;
mainly mc.cores when parallel::mclapply is chosen}
}
\description{
Read data from a flowjo wsp file
}
