% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wsx_get_popstats.R
\name{wsx_get_popstats}
\alias{wsx_get_popstats}
\title{Convenient function to import population counts from a flowjo workspace}
\usage{
wsx_get_popstats(ws, return_stats = T)
}
\arguments{
\item{ws}{path to flowjo workspace or a parsed xml-document (xml2::read_xml(ws))}

\item{return_stats}{return statistics next to cells counts}
}
\value{
returns a data.frame with cells counts or a list with counts and statistics if return_stats = T
}
\description{
In comparison to ws_get_popstats another backend (xml2) is used to read data from a flowjo .wsp-file (under the hood an .xml-file).
No groups or similar have to be selected - filtering has to be conducted afterwards.
Also, no restriction with respect to inconsistent gating trees are made. Every sample in handled individually.
All information are gathered from the .wsp file, no FCS files are touched or required.
}
\examples{
\dontrun{
# When the script is saved to R_scripts in the experiment folder,
# get the absolute path to the folder
wd <- dirname(dirname(rstudioapi::getActiveDocumentContext()$path))
# find workspaces
ws <- list.files(path = wd, pattern = '\\\\.wsp$', recursive = T, full.names = T)
# import the population counts:
wsx_get_popstats(ws = ws[1])
}
}
