% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wsx_get_popstats.R
\name{wsx_get_popstats}
\alias{wsx_get_popstats}
\title{Convenient function to import population counts from a flowjo workspace}
\usage{
wsx_get_popstats(ws)
}
\arguments{
\item{ws}{character of path to flowjo workspace}
}
\value{
returns a data.frame with population counts
}
\description{
In comparison to ws_get_popstats another backend (xml2) is used to read data from a flowjo .wsp-file (under the hood an .xml-file).
As is should be quick to extract data no groups have to be selected. Filtering has to be conducted afterwards.
Also, no restriction to gating trees within a group are made and no folder of FCS has to be provided.
}
\examples{
\dontrun{
# When the script is saved to R_scripts in the experiment folder,
# get the absolute path to the folder
wd <- dirname(dirname(rstudioapi::getActiveDocumentContext()$path))
# find workspaces
ws <- list.files(path = wd, pattern = '\\\\.wsp$', recursive = T, full.names = T)
# import the population counts:
wsx_get_popstats(ws = ws[1])
}
}
