% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wsx_get_popstats.R
\name{wsx_get_popstats}
\alias{wsx_get_popstats}
\title{Import data from a wsp file}
\usage{
wsx_get_popstats(ws, return_stats = T)
}
\arguments{
\item{ws}{path to flowjo workspace or a parsed xml-document (xml2::read_xml(ws))}

\item{return_stats}{return statistics next to cells counts}
}
\value{
data frame with cells counts or a list with counts and statistics if return_stats = T
}
\description{
Flowjo wsp files contain many information like gated event counts, statistics and keywords
from FCS files. These may be accessed without a dongle and can be read completely independent of
the respective FCS files once the gating has been conducted.
}
\examples{
\dontrun{
# When the script is saved to R_scripts in the experiment folder,
# get the absolute path to the folder
wd <- dirname(dirname(rstudioapi::getActiveDocumentContext()$path))
# find workspaces
ws <- list.files(path = wd, pattern = '\\\\.wsp$', recursive = T, full.names = T)
# import the population counts:
wsx_get_popstats(ws = ws[[1]])
}
}
