% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wsx_get_popstats_legacy.R
\name{wsx_get_popstats_legacy}
\alias{wsx_get_popstats_legacy}
\title{Import data from a wsp file}
\usage{
wsx_get_popstats_legacy(
  ws,
  groups = NULL,
  invert_groups = F,
  return_stats = F,
  lapply_fun = lapply,
  strip_data = T,
  ...
)
}
\arguments{
\item{ws}{path to flowjo workspace or a parsed xml-document (xml2::read_xml(ws))}

\item{groups}{vector of flowjo group names to consider}

\item{invert_groups}{logical whether to exclude the selected groups}

\item{return_stats}{logical whether to return statistics}

\item{lapply_fun}{function name without quotes; lapply, pbapply::pblapply or parallel::mclapply are suggested}

\item{...}{additional argument to the lapply function; mainly mc.cores when parallel::mclapply is chosen}
}
\value{
list of data frames: counts and optionally stats
}
\description{
Flowjo wsp files contain many information like gated event counts, statistics and keywords
from FCS files. These may be accessed without a dongle and can be read completely independent of
the respective FCS files once the gating has been conducted.
}
\examples{
\dontrun{
# find workspaces
ws <- list.files(path = wd, pattern = '\\\\.wsp$', recursive = T, full.names = T)
# read counts
counts <- wsx_get_popstats_legacy(ws = ws[[1]])[["counts"]]
}
}
